refactor(api): split handler.go into focused handler files
