+++ internal/api/handler.go
@@ -1,20 +1,11 @@
 package api
 
 import (
-	"context"
-	"encoding/base64"
 	"encoding/json"
-	"fmt"
-	"io"
 	"log/slog"
 	"net"
 	"net/http"
 	"net/url"
-	"os"
-	"path/filepath"
-	"runtime"
-	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -24,19 +15,14 @@ import (
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/hls"
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/merge"
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/middleware"
-	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/model"
-	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/onvif"
-	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/recorder"
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/storage"
 	"github.com/go-chi/chi/v5"
 )
 
 var logger = slog.Default().With("component", "api")
 
-
-
-
 var appStartTime = time.Now()
+
 // HealthCheck represents the result of a single health check.
 type HealthCheck struct {
 	Status  string `json:"status"` // "ok" | "warning" | "error"
@@ -186,1420 +172,6 @@ func (h *Handler) Routes() http.Handler {
 	return r
 }
 
-// --- Public endpoints ---
-
-func (h *Handler) handleHealth(w http.ResponseWriter, r *http.Request) {
-	resp := HealthResponse{Checks: make(map[string]HealthCheck)}
-	hasWarning, hasError := false, false
-
-	// Database check
-	if h.db != nil {
-		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
-		defer cancel()
-		err := h.db.DB().PingContext(ctx)
-		if err != nil {
-			resp.Checks["database"] = HealthCheck{Status: "error", Message: err.Error()}
-			hasError = true
-		} else {
-			resp.Checks["database"] = HealthCheck{Status: "ok"}
-		}
-	} else {
-		resp.Checks["database"] = HealthCheck{Status: "error", Message: "database not configured"}
-		hasError = true
-	}
-
-	// Storage check
-	if h.store != nil {
-		total, used, err := h.store.GetDiskUsage()
-		if err != nil {
-			resp.Checks["storage"] = HealthCheck{Status: "error", Message: err.Error()}
-			hasError = true
-		} else {
-			pct := 0
-			if total > 0 {
-				pct = int(float64(used) / float64(total) * 100)
-			}
-			msg := fmt.Sprintf("%d%% used (%d / %d bytes)", pct, used, total)
-			if pct > 95 {
-				resp.Checks["storage"] = HealthCheck{Status: "error", Message: msg}
-				hasError = true
-			} else if pct > 90 {
-				resp.Checks["storage"] = HealthCheck{Status: "warning", Message: msg}
-				hasWarning = true
-			} else {
-				resp.Checks["storage"] = HealthCheck{Status: "ok", Message: msg}
-			}
-		}
-	} else {
-		resp.Checks["storage"] = HealthCheck{Status: "error", Message: "storage not configured"}
-		hasError = true
-	}
-
-	// Goroutine check
-	numGoroutines := runtime.NumGoroutine()
-	if numGoroutines > 1000 {
-		resp.Checks["goroutines"] = HealthCheck{Status: "error", Message: fmt.Sprintf("%d goroutines (threshold: 1000)", numGoroutines)}
-		hasError = true
-	} else {
-		resp.Checks["goroutines"] = HealthCheck{Status: "ok", Message: fmt.Sprintf("%d goroutines", numGoroutines)}
-	}
-
-	// Overall status
-	switch {
-	case hasError:
-		resp.Status = "unhealthy"
-	case hasWarning:
-		resp.Status = "degraded"
-	default:
-		resp.Status = "ok"
-	}
-
-	// Uptime
-	resp.Uptime = formatUptime(time.Since(appStartTime))
-
-	writeJSON(w, http.StatusOK, resp)
-}
-
-func (h *Handler) handleReadyz(w http.ResponseWriter, r *http.Request) {
-	checks := make(map[string]HealthCheck)
-
-	// Database must be ok
-	allOK := true
-	if h.db == nil {
-		checks["database"] = HealthCheck{Status: "error", Message: "database not configured"}
-		allOK = false
-	} else {
-		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
-		defer cancel()
-		if err := h.db.DB().PingContext(ctx); err != nil {
-			checks["database"] = HealthCheck{Status: "error", Message: err.Error()}
-			allOK = false
-		} else {
-			checks["database"] = HealthCheck{Status: "ok"}
-		}
-	}
-
-	// Storage must be < 95%
-	if h.store == nil {
-		checks["storage"] = HealthCheck{Status: "error", Message: "storage not configured"}
-		allOK = false
-	} else {
-		total, used, err := h.store.GetDiskUsage()
-		if err != nil {
-			checks["storage"] = HealthCheck{Status: "error", Message: err.Error()}
-			allOK = false
-		} else {
-			pct := 0
-			if total > 0 {
-				pct = int(float64(used) / float64(total) * 100)
-			}
-			if pct >= 95 {
-				checks["storage"] = HealthCheck{Status: "error", Message: fmt.Sprintf("%d%% used (threshold: 95%%)", pct)}
-				allOK = false
-			} else {
-				checks["storage"] = HealthCheck{Status: "ok"}
-			}
-		}
-	}
-
-	// Goroutines must be < 5000
-	numGoroutines := runtime.NumGoroutine()
-	if numGoroutines >= 5000 {
-		checks["goroutines"] = HealthCheck{Status: "error", Message: fmt.Sprintf("%d goroutines (threshold: 5000)", numGoroutines)}
-		allOK = false
-	} else {
-		checks["goroutines"] = HealthCheck{Status: "ok"}
-	}
-
-	if allOK {
-		writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-	} else {
-		writeJSON(w, http.StatusServiceUnavailable, map[string]any{"status": "not ready", "checks": checks})
-	}
-}
-
-func (h *Handler) handleLogin(w http.ResponseWriter, r *http.Request) {
-	// Validate credentials by running through the auth middleware.
-	// If auth is disabled, any request succeeds; otherwise BasicAuth is checked.
-	done := make(chan int, 1)
-
-	inner := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		done <- http.StatusOK
-	})
-	rec := &middleware.StatusRecorder{ResponseWriter: w, Status: http.StatusUnauthorized}
-	h.authMW(inner).ServeHTTP(rec, r)
-
-	select {
-	case status := <-done:
-		if status == http.StatusOK {
-			writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-		}
-	default:
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusUnauthorized)
-		json.NewEncoder(w).Encode(map[string]string{"error": "unauthorized"})
-	}
-}
-
-// --- Recording endpoints ---
-
-func (h *Handler) handleListRecordings(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	filter := model.RecordingFilter{
-		CameraID: r.URL.Query().Get("camera_id"),
-		Format:   model.Format(r.URL.Query().Get("format")),
-	}
-
-	if v := r.URL.Query().Get("merged"); v != "" {
-		merged := v == "true" || v == "1"
-		filter.Merged = &merged
-	}
-
-	if v := r.URL.Query().Get("start"); v != "" {
-		if t, err := time.Parse(time.RFC3339, v); err == nil {
-			filter.StartTime = t
-		}
-	}
-
-	if v := r.URL.Query().Get("end"); v != "" {
-		if t, err := time.Parse(time.RFC3339, v); err == nil {
-			filter.EndTime = t
-		}
-	}
-
-	if v := r.URL.Query().Get("limit"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil && n > 0 {
-			filter.Limit = n
-		}
-	}
-
-	if v := r.URL.Query().Get("offset"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil && n >= 0 {
-			filter.Offset = n
-		}
-	}
-
-	// Sorting
-	filter.SortBy = r.URL.Query().Get("sort_by")
-	filter.SortOrder = r.URL.Query().Get("order")
-
-	filter.Search = r.URL.Query().Get("search")
-
-	recordings, err := h.db.ListRecordings(ctx, filter)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to list recordings")
-		return
-	}
-
-	if recordings == nil {
-		recordings = []model.Recording{}
-	}
-
-	total, err := h.db.CountRecordingsWithFilter(ctx, filter)
-	if err != nil {
-		total = 0 // non-fatal
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"recordings": recordings,
-		"total":      total,
-	})
-}
-
-func (h *Handler) handleGetRecording(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	rec, err := h.db.GetRecording(r.Context(), id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get recording")
-		return
-	}
-	if rec == nil {
-		writeError(w, http.StatusNotFound, "recording not found")
-		return
-	}
-	writeJSON(w, http.StatusOK, rec)
-}
-
-func (h *Handler) handleDeleteRecording(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	ctx := r.Context()
-
-	rec, err := h.db.GetRecording(ctx, id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get recording")
-		return
-	}
-	if rec == nil {
-		writeError(w, http.StatusNotFound, "recording not found")
-		return
-	}
-
-	// Delete from DB first (authoritative source)
-	if err := h.db.DeleteRecording(ctx, id); err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to delete recording")
-		return
-	}
-
-	// Then delete file (non-fatal if fails)
-	if rec.FilePath != "" {
-		if err := h.store.DeleteFile(rec.FilePath); err != nil {
-			logger.Warn("failed to delete file", "file_path", rec.FilePath, "error", err)
-		}
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "deleted"})
-}
-
-func (h *Handler) handleBatchDeleteRecordings(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	var body struct {
-		IDs []string `json:"ids"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if len(body.IDs) == 0 {
-		writeError(w, http.StatusBadRequest, "ids must not be empty")
-		return
-	}
-	if len(body.IDs) > 100 {
-		writeError(w, http.StatusBadRequest, "ids must not exceed 100")
-		return
-	}
-	// Fetch file paths before batch delete
-	filePaths := map[string]string{}
-	for _, id := range body.IDs {
-		rec, err := h.db.GetRecording(ctx, id)
-		if err == nil && rec != nil && rec.FilePath != "" {
-			filePaths[id] = rec.FilePath
-		}
-	}
-
-	// Delete DB records (transaction)
-	deleted, err := h.db.DeleteRecordingsBatch(ctx, body.IDs)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to delete recordings")
-		return
-	}
-
-	// Attempt file deletion for successfully deleted records (non-fatal)
-	failed := []string{}
-	deletedSet := make(map[string]bool, len(deleted))
-	for _, id := range deleted {
-		deletedSet[id] = true
-		if fp, ok := filePaths[id]; ok {
-			if err := h.store.DeleteFile(fp); err != nil {
-				logger.Warn("batch delete: failed to delete file", "file_path", fp, "error", err)
-			}
-		}
-	}
-	for _, id := range body.IDs {
-		if !deletedSet[id] {
-			failed = append(failed, id)
-		}
-	}
-
-	result := map[string]any{"deleted": deleted}
-	if len(failed) > 0 {
-		result["failed"] = failed
-	} else {
-		result["failed"] = []string{}
-	}
-	writeJSON(w, http.StatusOK, result)
-}
-
-func (h *Handler) handleDownloadRecording(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	rec, err := h.db.GetRecording(r.Context(), id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get recording")
-		return
-	}
-	if rec == nil {
-		writeError(w, http.StatusNotFound, "recording not found")
-		return
-	}
-
-	if rec.FilePath == "" {
-		writeError(w, http.StatusNotFound, "file not available")
-		return
-	}
-
-	// Validate that the recording file path is within the storage root to prevent
-	// path traversal. This ensures rec.FilePath (which may come from external
-	// sources like WebDAV uploads) is confined to the storage directory.
-	validPath, err := storage.ValidatePath(h.store.RootDir(), rec.FilePath)
-	if err != nil {
-		writeError(w, http.StatusNotFound, "invalid file path")
-		return
-	}
-
-	// Check for frame parameter (MJPEG frame download)
-	frameStr := r.URL.Query().Get("frame")
-	if frameStr != "" && rec.Format == model.FormatMJPEG {
-		frameIndex, err := strconv.Atoi(frameStr)
-		if err == nil {
-			entries, err := os.ReadDir(validPath)
-			if err == nil {
-				jpgFiles := []os.DirEntry{}
-				for _, e := range entries {
-					if !e.IsDir() && isImageFile(e.Name()) {
-						jpgFiles = append(jpgFiles, e)
-					}
-				}
-				sort.Slice(jpgFiles, func(i, j int) bool { return jpgFiles[i].Name() < jpgFiles[j].Name() })
-				if frameIndex >= 0 && frameIndex < len(jpgFiles) {
-					framePath := filepath.Join(validPath, jpgFiles[frameIndex].Name())
-					http.ServeFile(w, r, framePath)
-					return
-				}
-			}
-		}
-		http.Error(w, "frame not found", http.StatusNotFound)
-		return
-	}
-
-	filePath := validPath
-	info, err := os.Stat(filePath)
-	if err != nil {
-		writeError(w, http.StatusNotFound, "file not found")
-		return
-	}
-	if info.IsDir() {
-		entries, err := os.ReadDir(filePath)
-		if err != nil || len(entries) == 0 {
-			writeError(w, http.StatusNotFound, "no files in recording directory")
-			return
-		}
-		for _, e := range entries {
-			if e.IsDir() {
-				continue
-			}
-			name := e.Name()
-			if strings.HasSuffix(name, ".jpg") || strings.HasSuffix(name, ".jpeg") || strings.HasSuffix(name, ".mp4") {
-				filePath = filepath.Join(filePath, name)
-				break
-			}
-		}
-	}
-
-	w.Header().Set("Content-Disposition", fmt.Sprintf("inline; filename=\"%s\"", filepath.Base(filePath)))
-	http.ServeFile(w, r, filePath)
-}
-
-func (h *Handler) handleListFrames(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	rec, err := h.db.GetRecording(r.Context(), id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get recording")
-		return
-	}
-	if rec == nil {
-		writeError(w, http.StatusNotFound, "recording not found")
-		return
-	}
-
-	if rec.Format != "mjpeg" {
-		writeError(w, http.StatusBadRequest, "not a JPEG recording")
-		return
-	}
-
-	filePath := rec.FilePath
-	info, err := os.Stat(filePath)
-	if err != nil {
-		writeError(w, http.StatusNotFound, "recording files not found")
-		return
-	}
-	if !info.IsDir() {
-		writeError(w, http.StatusNotFound, "recording is not a directory")
-		return
-	}
-
-	entries, err := os.ReadDir(filePath)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to read recording directory")
-		return
-	}
-
-	type FrameInfo struct {
-		Index    int    `json:"index"`
-		Filename string `json:"filename"`
-		Size     int64  `json:"size"`
-	}
-
-	var frames []FrameInfo
-	for _, e := range entries {
-		if e.IsDir() {
-			continue
-		}
-		name := e.Name()
-		if !strings.HasSuffix(strings.ToLower(name), ".jpg") && !strings.HasSuffix(strings.ToLower(name), ".jpeg") {
-			continue
-		}
-		fi, err := e.Info()
-		if err != nil {
-			continue
-		}
-		frames = append(frames, FrameInfo{
-			Filename: name,
-			Size:     fi.Size(),
-		})
-	}
-
-	// Sort by filename (natural order - timestamp-based names sort correctly)
-	sort.Slice(frames, func(i, j int) bool {
-		return frames[i].Filename < frames[j].Filename
-	})
-
-	// Assign sequential indices
-	for i := range frames {
-		frames[i].Index = i
-	}
-
-	if frames == nil {
-		frames = []FrameInfo{}
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"frames": frames,
-	})
-}
-
-// --- Camera and stats endpoints ---
-
-// cameraRowForAPI normalizes camera rows for API responses.
-// For ONVIF cameras, it exposes onvif_endpoint as url so the frontend
-// can use a single url field for all protocols.
-func cameraRowForAPI(row *storage.CameraRow) {
-	if row.Protocol == "onvif" && row.URL == "" && row.ONVIFEndpoint != "" {
-		row.URL = row.ONVIFEndpoint
-	}
-}
-
-func (h *Handler) handleListCameras(w http.ResponseWriter, r *http.Request) {
-	cameras, err := h.db.ListCameras(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to list cameras")
-		return
-	}
-	if cameras == nil {
-		cameras = []storage.CameraRow{}
-	}
-	// Inject recorder status from CameraManager
-	if h.camMgr != nil {
-		statusMap := h.camMgr.Status()
-		for i := range cameras {
-			if s, ok := statusMap[cameras[i].ID]; ok {
-				cameras[i].Status = s
-			} else {
-				cameras[i].Status = model.StatusStopped
-			}
-		}
-		// Inject last_seen from DB
-		lastSeenMap, err := h.db.GetAllLastRecordingTimes(r.Context())
-		if err == nil {
-			for i := range cameras {
-				if t, ok := lastSeenMap[cameras[i].ID]; ok {
-					cameras[i].LastSeen = t
-				}
-			}
-		}
-	}
-	// For ONVIF cameras, show onvif_endpoint as url for unified frontend handling
-	for i := range cameras {
-		cameraRowForAPI(&cameras[i])
-	}
-	writeJSON(w, http.StatusOK, cameras)
-}
-
-func (h *Handler) handleStats(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	total, used, err := h.store.GetDiskUsage()
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get disk usage")
-		return
-	}
-
-	count, err := h.db.CountRecordings(ctx)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to count recordings")
-		return
-	}
-
-	cameras, err := h.db.ListCameras(ctx)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to count cameras")
-		return
-	}
-
-	stats := model.StorageStats{
-		TotalBytes:     total,
-		UsedBytes:      used,
-		RecordingCount: count,
-		CameraCount:    len(cameras),
-	}
-	writeJSON(w, http.StatusOK, stats)
-}
-
-func (h *Handler) handleStatsTrends(w http.ResponseWriter, r *http.Request) {
-	days := 7
-	if v := r.URL.Query().Get("days"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil && n >= 1 && n <= 30 {
-			days = n
-		}
-	}
-	trends, err := h.db.GetRecordingTrends(r.Context(), days)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get recording trends")
-		return
-	}
-	writeJSON(w, http.StatusOK, trends)
-}
-
-// --- Camera CRUD endpoints ---
-
-var validProtocols = map[string]bool{
-	// New transport-only protocols
-	"rtsp":  true,
-	"http":  true,
-	"onvif": true,
-	// Plugin protocols
-	"xiaomi": true,
-	// Legacy combined protocols (accepted, will be normalized)
-	"rtsp_h264":  true,
-	"rtsp_h265":  true,
-	"rtsp_mjpeg": true,
-	"http_jpeg":  true,
-}
-
-func (h *Handler) handleCreateCamera(w http.ResponseWriter, r *http.Request) {
-	var body struct {
-		Name           string `json:"name"`
-		Protocol       string `json:"protocol"`
-		URL            string `json:"url"`
-		Username       string `json:"username"`
-		Password       string `json:"password"`
-		Enabled        *bool  `json:"enabled"`
-		Description    string `json:"description"`
-		Location       string `json:"location"`
-		Brand          string `json:"brand"`
-		Model          string `json:"model"`
-		SerialNumber   string `json:"serial_number"`
-		ONVIFEndpoint  string `json:"onvif_endpoint"`
-		ProfileToken   string `json:"profile_token"`
-		StreamEncoding string `json:"stream_encoding"`
-		Encoding       string `json:"encoding"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if body.Name == "" {
-		writeError(w, http.StatusBadRequest, "name is required")
-		return
-	}
-	if body.Protocol == "" {
-		writeError(w, http.StatusBadRequest, "protocol is required")
-		return
-	}
-	if !validProtocols[body.Protocol] {
-		writeError(w, http.StatusBadRequest, fmt.Sprintf("invalid protocol %q, must be one of: rtsp, http, onvif, xiaomi", body.Protocol))
-		return
-	}
-	// ONVIF cameras: accept url OR onvif_endpoint
-	if body.Protocol == "onvif" {
-		endpoint := body.ONVIFEndpoint
-		if endpoint == "" {
-			endpoint = body.URL
-		}
-		if endpoint == "" {
-			writeError(w, http.StatusBadRequest, "url or onvif_endpoint is required for ONVIF cameras")
-			return
-		}
-		body.ONVIFEndpoint = endpoint
-		body.URL = "" // Don't store in url field for ONVIF
-		// Check for duplicate ONVIF endpoint
-		if h.db != nil {
-			existingCams, _ := h.db.ListCameras(r.Context())
-			for _, ec := range existingCams {
-				if ec.Protocol == "onvif" && ec.ONVIFEndpoint == body.ONVIFEndpoint {
-					writeError(w, http.StatusConflict, "ONVIF camera with this endpoint already exists")
-					return
-				}
-			}
-		}
-	} else if body.URL == "" {
-		writeError(w, http.StatusBadRequest, "url is required")
-		return
-	}
-	// Validate URL format for non-ONVIF cameras
-	if body.Protocol != "onvif" && !validateURL(body.URL) {
-		writeError(w, http.StatusBadRequest, "invalid URL format")
-		return
-	}
-	// Normalize protocol — handle legacy combined formats
-	proto := body.Protocol
-	enc := body.Encoding
-	if strings.Contains(proto, "_") {
-		parsedProto, parsedEnc, err := model.ParseLegacyProtocol(proto)
-		if err != nil {
-			writeError(w, http.StatusBadRequest, fmt.Sprintf("invalid protocol %q", proto))
-			return
-		}
-		proto = parsedProto
-		if enc == "" {
-			enc = parsedEnc
-		}
-	}
-	// Set default encoding if still empty
-	if enc == "" {
-		switch proto {
-		case "rtsp":
-			enc = "h264"
-		case "http":
-			enc = "jpeg"
-		}
-	}
-
-	cam := config.CameraConfig{
-		Name:           body.Name,
-		Protocol:       proto,
-		Encoding:       enc,
-		URL:            body.URL,
-		Username:       body.Username,
-		Password:       body.Password,
-		ONVIFEndpoint:  body.ONVIFEndpoint,
-		ProfileToken:   body.ProfileToken,
-		StreamEncoding: body.StreamEncoding,
-	}
-	if body.Enabled != nil {
-		cam.Enabled = *body.Enabled
-	} else {
-		cam.Enabled = true
-	}
-
-	if h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "camera manager not available")
-		return
-	}
-	id, err := h.camMgr.AddCamera(r.Context(), cam)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, fmt.Sprintf("failed to add camera: %v", err))
-		return
-	}
-	// Persist DB-only metadata fields
-	if body.Description != "" || body.Location != "" || body.Brand != "" || body.Model != "" || body.SerialNumber != "" {
-		if err := h.db.UpdateCameraMetadata(r.Context(), id, body.Description, body.Location, body.Brand, body.Model, body.SerialNumber, 0); err != nil {
-			logger.Warn("failed to set camera metadata", "camera_id", id, "error", err)
-		}
-	}
-	// Return CameraRow with status
-	row, err := h.db.GetCamera(r.Context(), id)
-	if row != nil {
-		if h.camMgr != nil {
-			row.Status = h.camMgr.CameraStatus(id)
-		}
-		// Inject last_seen from DB
-		lastSeen, err := h.db.GetLastRecordingTime(r.Context(), id)
-		if err == nil {
-			row.LastSeen = lastSeen
-		}
-		cameraRowForAPI(row)
-		writeJSON(w, http.StatusCreated, row)
-	} else {
-		cam.ID = id
-		writeJSON(w, http.StatusCreated, cam)
-	}
-}
-
-func (h *Handler) handleGetCamera(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	row, err := h.db.GetCamera(r.Context(), id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get camera")
-		return
-	}
-	if row == nil {
-		writeError(w, http.StatusNotFound, "camera not found")
-		return
-	}
-	// Inject recorder status
-	if h.camMgr != nil {
-		row.Status = h.camMgr.CameraStatus(id)
-	}
-	// Inject last_seen from DB
-	lastSeen, err := h.db.GetLastRecordingTime(r.Context(), id)
-	if err == nil {
-		row.LastSeen = lastSeen
-	}
-	cameraRowForAPI(row)
-	writeJSON(w, http.StatusOK, row)
-}
-
-func (h *Handler) handleUpdateCamera(w http.ResponseWriter, r *http.Request) {
-	if h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "camera manager not available")
-		return
-	}
-	id := chi.URLParam(r, "id")
-
-	var body struct {
-		Name           *string `json:"name"`
-		URL            *string `json:"url"`
-		Protocol       *string `json:"protocol"`
-		Encoding       *string `json:"encoding"`
-		Username       *string `json:"username"`
-		Password       *string `json:"password"`
-		Enabled        *bool   `json:"enabled"`
-		Description    *string `json:"description"`
-		Location       *string `json:"location"`
-		Brand          *string `json:"brand"`
-		Model          *string `json:"model"`
-		SerialNumber   *string `json:"serial_number"`
-		RetentionDays  *int    `json:"retention_days"`
-		ONVIFEndpoint  *string `json:"onvif_endpoint"`
-		ProfileToken   *string `json:"profile_token"`
-		StreamEncoding *string `json:"stream_encoding"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	// Harden credential updates: empty string from frontend means "don't update"
-	username := body.Username
-	if username != nil && *username == "" {
-		username = nil
-	}
-	password := body.Password
-	if password != nil && *password == "" {
-		password = nil
-	}
-
-	updates := camera.CameraUpdate{
-		Name:           body.Name,
-		URL:            body.URL,
-		Protocol:       body.Protocol,
-		Encoding:       body.Encoding,
-		Username:       username,
-		Password:       password,
-		Enabled:        body.Enabled,
-		Description:    body.Description,
-		Location:       body.Location,
-		Brand:          body.Brand,
-		Model:          body.Model,
-		SerialNumber:   body.SerialNumber,
-		RetentionDays:  body.RetentionDays,
-		ONVIFEndpoint:  body.ONVIFEndpoint,
-		ProfileToken:   body.ProfileToken,
-		StreamEncoding: body.StreamEncoding,
-	}
-
-	// Validate URL format if URL is being updated
-	if body.URL != nil && *body.URL != "" {
-		if body.Protocol == nil || *body.Protocol != "onvif" {
-			if !validateURL(*body.URL) {
-				writeError(w, http.StatusBadRequest, "invalid URL format")
-				return
-			}
-		}
-	}
-
-	// For ONVIF cameras, sync url and onvif_endpoint
-	if body.Protocol != nil && *body.Protocol == "onvif" {
-		if updates.URL != nil && *updates.URL != "" {
-			updates.ONVIFEndpoint = updates.URL
-			updates.URL = nil
-		}
-		if updates.ONVIFEndpoint != nil && *updates.ONVIFEndpoint != "" {
-			updates.URL = updates.ONVIFEndpoint
-		}
-	}
-
-	_, err := h.camMgr.UpdateCamera(r.Context(), id, updates)
-	if err != nil {
-		if strings.Contains(err.Error(), "not found") {
-			writeError(w, http.StatusNotFound, "camera not found")
-			return
-		}
-		writeError(w, http.StatusInternalServerError, fmt.Sprintf("failed to update camera: %v", err))
-		return
-	}
-	// Return updated CameraRow with status
-	row, err := h.db.GetCamera(r.Context(), id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get camera")
-		return
-	}
-	if row != nil {
-		if h.camMgr != nil {
-			row.Status = h.camMgr.CameraStatus(id)
-		}
-		// Inject last_seen from DB
-		lastSeen, err := h.db.GetLastRecordingTime(r.Context(), id)
-		if err == nil {
-			row.LastSeen = lastSeen
-		}
-		cameraRowForAPI(row)
-		writeJSON(w, http.StatusOK, row)
-	} else {
-		writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-	}
-}
-
-func (h *Handler) handleDeleteCamera(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	ctx := r.Context()
-
-	// Verify camera exists in DB
-	cam, err := h.db.GetCamera(ctx, id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get camera")
-		return
-	}
-	if cam == nil {
-		writeError(w, http.StatusNotFound, "camera not found")
-		return
-	}
-
-	// Archive the camera: stops recorder, merges segments, marks archived in DB, removes from config.
-	// This preserves the camera row and recordings for the archive view.
-	if h.camMgr != nil {
-		if err := h.camMgr.ArchiveCamera(ctx, id); err != nil {
-			// Camera may not be in config (orphaned DB-only) — archive directly in DB
-			logger.Warn("failed to archive camera via manager, archiving in DB", "camera_id", id, "error", err)
-			if dbErr := h.db.ArchiveCameraDB(ctx, id); dbErr != nil {
-				writeError(w, http.StatusInternalServerError, "failed to archive camera")
-				return
-			}
-			if _, recErr := h.db.ArchiveAllRecordings(ctx, id); recErr != nil {
-				logger.Warn("failed to archive recordings", "camera_id", id, "error", recErr)
-			}
-		}
-	} else {
-		// No camera manager — mark archived in DB directly
-		if err := h.db.ArchiveCameraDB(ctx, id); err != nil {
-			writeError(w, http.StatusInternalServerError, "failed to archive camera")
-			return
-		}
-		if _, err := h.db.ArchiveAllRecordings(ctx, id); err != nil {
-			logger.Warn("failed to archive recordings", "camera_id", id, "error", err)
-		}
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "archived"})
-}
-
-func (h *Handler) handleStartCamera(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	if h.camMgr == nil {
-		writeError(w, http.StatusServiceUnavailable, "camera manager not available")
-		return
-	}
-	if err := h.camMgr.StartCamera(r.Context(), id); err != nil {
-		errMsg := err.Error()
-		switch {
-		case strings.Contains(errMsg, "not found"):
-			writeError(w, http.StatusNotFound, errMsg)
-		case strings.Contains(errMsg, "disabled"):
-			writeError(w, http.StatusBadRequest, errMsg)
-		case strings.Contains(errMsg, "already running"):
-			writeError(w, http.StatusConflict, errMsg)
-		default:
-			writeError(w, http.StatusInternalServerError, errMsg)
-		}
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "started"})
-}
-
-func (h *Handler) handleStopCamera(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	if h.camMgr == nil {
-		writeError(w, http.StatusServiceUnavailable, "camera manager not available")
-		return
-	}
-	if err := h.camMgr.StopCamera(r.Context(), id); err != nil {
-		if strings.Contains(err.Error(), "not found") {
-			writeError(w, http.StatusNotFound, err.Error())
-		} else {
-			writeError(w, http.StatusInternalServerError, err.Error())
-		}
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "stopped"})
-}
-
-// --- Archive endpoints ---
-
-// archiveGroupItem is the JSON response for a single archive group.
-type archiveGroupItem struct {
-	ID                   string     `json:"id"`
-	Name                 string     `json:"name"`
-	RecordingCount       int        `json:"recording_count"`
-	TotalSize            int64      `json:"total_size"`
-	ArchivedAt           *time.Time `json:"archived_at,omitempty"`
-	ArchiveRetentionDays int        `json:"archive_retention_days"`
-}
-
-func (h *Handler) handleListArchives(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	cameras, err := h.db.ListArchivedCameras(ctx)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to list archived cameras")
-		return
-	}
-	if cameras == nil {
-		cameras = []storage.CameraRow{}
-	}
-
-	items := make([]archiveGroupItem, 0, len(cameras))
-	for _, cam := range cameras {
-		count, totalSize, err := h.db.GetArchiveGroupStats(ctx, cam.ID)
-		if err != nil {
-			logger.Warn("failed to get archive stats", "camera_id", cam.ID, "error", err)
-			count, totalSize = 0, 0
-		}
-		items = append(items, archiveGroupItem{
-			ID:                   cam.ID,
-			Name:                 cam.Name,
-			RecordingCount:       count,
-			TotalSize:            totalSize,
-			ArchivedAt:           cam.ArchivedAt,
-			ArchiveRetentionDays: cam.ArchiveRetentionDays,
-		})
-	}
-	writeJSON(w, http.StatusOK, map[string]interface{}{"archives": items})
-}
-
-func (h *Handler) handleListArchiveRecordings(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "cameraID")
-	ctx := r.Context()
-
-	trueVal := true
-	filter := model.RecordingFilter{
-		CameraID: cameraID,
-		Archived: &trueVal,
-	}
-	if v := r.URL.Query().Get("limit"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil && n > 0 {
-			filter.Limit = n
-		}
-	}
-	if v := r.URL.Query().Get("offset"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil && n >= 0 {
-			filter.Offset = n
-		}
-	}
-	filter.SortBy = r.URL.Query().Get("sort_by")
-	filter.SortOrder = r.URL.Query().Get("order")
-
-	recordings, err := h.db.ListRecordings(ctx, filter)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to list archived recordings")
-		return
-	}
-	if recordings == nil {
-		recordings = []model.Recording{}
-	}
-	total, err := h.db.CountRecordingsWithFilter(ctx, filter)
-	if err != nil {
-		total = 0
-	}
-	writeJSON(w, http.StatusOK, map[string]any{
-		"recordings": recordings,
-		"total":      total,
-	})
-}
-
-func (h *Handler) handleDeleteArchiveGroup(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "cameraID")
-	ctx := r.Context()
-
-	// Verify camera is archived
-	cam, err := h.db.GetCamera(ctx, cameraID)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get camera")
-		return
-	}
-	if cam == nil || !cam.Archived {
-		writeError(w, http.StatusNotFound, "archived camera not found")
-		return
-	}
-
-	// Get all archived recordings to delete files from disk
-	trueVal := true
-	recordings, err := h.db.ListRecordings(ctx, model.RecordingFilter{CameraID: cameraID, Archived: &trueVal})
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to list recordings")
-		return
-	}
-
-	// Delete recording files from disk (non-fatal)
-	for _, rec := range recordings {
-		if rec.FilePath != "" {
-			if err := h.store.DeleteFile(rec.FilePath); err != nil {
-				logger.Warn("failed to delete archived recording file", "file_path", rec.FilePath, "error", err)
-			}
-		}
-	}
-
-	// Delete all recordings from DB
-	ids := make([]string, len(recordings))
-	for i, rec := range recordings {
-		ids[i] = rec.ID
-	}
-	if len(ids) > 0 {
-		if _, err := h.db.DeleteRecordingsBatch(ctx, ids); err != nil {
-			writeError(w, http.StatusInternalServerError, "failed to delete recordings")
-			return
-		}
-	}
-
-	// Delete camera row from DB
-	if err := h.db.DeleteCamera(ctx, cameraID); err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to delete camera")
-		return
-	}
-
-	// Remove camera directory from disk
-	if err := h.store.DeleteCameraDir(cameraID); err != nil {
-		logger.Warn("failed to remove camera directory", "camera_id", cameraID, "error", err)
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "deleted"})
-}
-
-func (h *Handler) handleDeleteArchiveRecording(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "cameraID")
-	recordingID := chi.URLParam(r, "recordingID")
-	ctx := r.Context()
-
-	// Get the recording
-	rec, err := h.db.GetRecording(ctx, recordingID)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get recording")
-		return
-	}
-	if rec == nil || !rec.Archived || rec.CameraID != cameraID {
-		writeError(w, http.StatusNotFound, "archived recording not found")
-		return
-	}
-
-	// Delete file from disk (non-fatal)
-	if rec.FilePath != "" {
-		if err := h.store.DeleteFile(rec.FilePath); err != nil {
-			logger.Warn("failed to delete archived recording file", "file_path", rec.FilePath, "error", err)
-		}
-	}
-
-	// Delete recording from DB
-	if err := h.db.DeleteRecording(ctx, recordingID); err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to delete recording")
-		return
-	}
-
-	// Check if this was the last archived recording for this camera
-	count, _, err := h.db.GetArchiveGroupStats(ctx, cameraID)
-	if err == nil && count == 0 {
-		// No more recordings — clean up camera row and directory
-		if err := h.db.DeleteCamera(ctx, cameraID); err != nil {
-			logger.Warn("failed to delete empty archive camera", "camera_id", cameraID, "error", err)
-		}
-		if err := h.store.DeleteCameraDir(cameraID); err != nil {
-			logger.Warn("failed to remove camera directory", "camera_id", cameraID, "error", err)
-		}
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "deleted"})
-}
-
-func (h *Handler) handleSetArchiveRetention(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "cameraID")
-	ctx := r.Context()
-
-	var body struct {
-		RetentionDays int `json:"retention_days"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if body.RetentionDays < 0 {
-		writeError(w, http.StatusBadRequest, "retention_days must be >= 0")
-		return
-	}
-
-	if err := h.db.SetArchiveRetention(ctx, cameraID, body.RetentionDays); err != nil {
-		writeError(w, http.StatusNotFound, "archived camera not found")
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-// --- ONVIF camera management endpoints ---
-
-func (h *Handler) handleONVIFCameraProfiles(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "id")
-	if cameraID == "" {
-		writeError(w, http.StatusBadRequest, "camera ID is required")
-		return
-	}
-
-	// For now, return empty profiles (actual implementation needs ONVIF client)
-	writeJSON(w, http.StatusOK, map[string]interface{}{
-		"profiles":     []interface{}{},
-		"capabilities": map[string]bool{"ptz": false, "streaming": false},
-	})
-}
-
-// --- ONVIF discovery endpoints ---
-
-func (h *Handler) handleONVIFDiscover(w http.ResponseWriter, r *http.Request) {
-	var req struct {
-		Timeout int `json:"timeout"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		req.Timeout = 5
-	}
-	if req.Timeout <= 0 {
-		req.Timeout = 5
-	}
-	if req.Timeout > 30 {
-		writeError(w, http.StatusBadRequest, "timeout must be between 1 and 30 seconds")
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(r.Context(), time.Duration(req.Timeout)*time.Second)
-	defer cancel()
-
-	devices, err := onvif.Discover(ctx, time.Duration(req.Timeout)*time.Second)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, fmt.Sprintf("discovery failed: %v", err))
-		return
-	}
-	if devices == nil {
-		devices = []onvif.DiscoveredDevice{}
-	}
-	writeJSON(w, http.StatusOK, map[string]interface{}{"devices": devices})
-}
-
-func (h *Handler) handleONVIFDeviceDetail(w http.ResponseWriter, r *http.Request) {
-	ip := chi.URLParam(r, "ip")
-	if ip == "" {
-		writeError(w, http.StatusBadRequest, "IP address is required")
-		return
-	}
-	if !validateIP(ip) {
-		writeError(w, http.StatusBadRequest, "invalid IP address format")
-		return
-	}
-	ctx := r.Context()
-	client := onvif.NewClient(fmt.Sprintf("http://%s/onvif/device_service", ip), "", "")
-	if err := client.Connect(ctx); err != nil {
-		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to connect to device: %v", err))
-		return
-	}
-	info, err := client.GetDeviceInformation(ctx)
-	if err != nil {
-		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to get device info: %v", err))
-		return
-	}
-	profiles, err := client.GetProfiles(ctx)
-	if err != nil {
-		profiles = nil
-	}
-	writeJSON(w, http.StatusOK, map[string]interface{}{
-		"device_info": info,
-		"profiles":    profiles,
-	})
-}
-
-func (h *Handler) requireONVIF(w http.ResponseWriter, r *http.Request) bool {
-	if h.db == nil {
-		writeError(w, http.StatusNotFound, "camera not found")
-		return false
-	}
-	cameraID := chi.URLParam(r, "id")
-	camera, err := h.db.GetCamera(r.Context(), cameraID)
-	if err != nil || camera == nil {
-		writeError(w, http.StatusNotFound, "camera not found")
-		return false
-	}
-	if camera.Protocol != "onvif" {
-		writeError(w, http.StatusBadRequest, "PTZ control is only available for ONVIF cameras")
-		return false
-	}
-	return true
-}
-
-// --- PTZ control endpoints ---
-
-func (h *Handler) handlePTZMove(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "id")
-	var req struct {
-		Mode string  `json:"mode"`
-		Pan  float64 `json:"pan"`
-		Tilt float64 `json:"tilt"`
-		Zoom float64 `json:"zoom"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if req.Mode != "continuous" && req.Mode != "absolute" && req.Mode != "relative" {
-		writeError(w, http.StatusBadRequest, "mode must be continuous, absolute, or relative")
-		return
-	}
-	if !h.requireONVIF(w, r) {
-		return
-	}
-	if h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "camera manager not available")
-		return
-	}
-	ptz, err := h.camMgr.GetONVIFPTZController(r.Context(), cameraID)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	vec := onvif.PTZVector{Pan: req.Pan, Tilt: req.Tilt, Zoom: req.Zoom}
-	switch req.Mode {
-	case "continuous":
-		err = ptz.ContinuousMove(r.Context(), vec)
-	case "absolute":
-		err = ptz.AbsoluteMove(r.Context(), vec)
-	case "relative":
-		err = ptz.RelativeMove(r.Context(), vec)
-	}
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, fmt.Sprintf("PTZ command failed: %v", err))
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-}
-
-func (h *Handler) handlePTZStop(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "id")
-	if !h.requireONVIF(w, r) {
-		return
-	}
-	if h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "camera manager not available")
-		return
-	}
-	ptz, err := h.camMgr.GetONVIFPTZController(r.Context(), cameraID)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	if err := ptz.Stop(r.Context(), true, true); err != nil {
-		writeError(w, http.StatusInternalServerError, fmt.Sprintf("PTZ stop failed: %v", err))
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "stopped"})
-}
-
-func (h *Handler) handlePTZStatus(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "id")
-	if !h.requireONVIF(w, r) {
-		return
-	}
-	if h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "camera manager not available")
-		return
-	}
-	ptz, err := h.camMgr.GetONVIFPTZController(r.Context(), cameraID)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	pos, moving, err := ptz.GetStatus(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, fmt.Sprintf("get PTZ status failed: %v", err))
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]interface{}{
-		"pan":    pos.Pan,
-		"tilt":   pos.Tilt,
-		"zoom":   pos.Zoom,
-		"moving": moving,
-	})
-}
-
-// --- Snapshot endpoint ---
-
-func (h *Handler) handleSnapshot(w http.ResponseWriter, r *http.Request) {
-	cameraID := chi.URLParam(r, "id")
-
-	// Find camera config to get SnapshotURL
-	var snapshotURL string
-	if h.config != nil {
-		for _, cam := range h.config.Cameras {
-			if cam.ID == cameraID {
-				snapshotURL = cam.SnapshotURL
-				break
-			}
-		}
-	}
-	if snapshotURL == "" {
-		http.Error(w, "Snapshot URL not configured", http.StatusNotFound)
-		return
-	}
-
-	// Check cache (10 second TTL)
-	const cacheTTL = 10 * time.Second
-	h.snapshotMu.RLock()
-	cached, ok := h.snapshots[cameraID]
-	h.snapshotMu.RUnlock()
-
-	if ok && time.Since(cached.timestamp) < cacheTTL {
-		w.Header().Set("Content-Type", "image/jpeg")
-		w.Header().Set("Cache-Control", "max-age=5")
-		w.Write(cached.data)
-		return
-	}
-
-	// Fetch from camera
-	client := &http.Client{Timeout: 5 * time.Second}
-	resp, err := client.Get(snapshotURL)
-	if err != nil {
-		// Return stale cache if available
-		if ok {
-			w.Header().Set("Content-Type", "image/jpeg")
-			w.Header().Set("X-Cache", "stale")
-			w.Write(cached.data)
-			return
-		}
-		logger.Warn("failed to fetch snapshot", "camera_id", cameraID, "error", err)
-		http.Error(w, "Failed to fetch snapshot", http.StatusBadGateway)
-		return
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		http.Error(w, "Camera returned error", http.StatusBadGateway)
-		return
-	}
-
-	data, err := io.ReadAll(io.LimitReader(resp.Body, 10*1024*1024)) // 10MB max
-	if err != nil || len(data) == 0 {
-		http.Error(w, "Failed to read snapshot", http.StatusBadGateway)
-		return
-	}
-
-	// Update cache
-	h.snapshotMu.Lock()
-	h.snapshots[cameraID] = &snapshotCache{data: data, timestamp: time.Now()}
-	h.snapshotMu.Unlock()
-
-	w.Header().Set("Content-Type", "image/jpeg")
-	w.Header().Set("Cache-Control", "max-age=5")
-	w.Write(data)
-}
-
 // --- Helpers ---
 
 func writeJSON(w http.ResponseWriter, status int, v any) {
@@ -1662,1049 +234,6 @@ func TestHandler(db *storage.DB, store *storage.Manager) *Handler {
 func TestHandlerWithAuth(db *storage.DB, store *storage.Manager, username, passwordHash string) *Handler {
 	authMW, _ := middleware.NewAuthMiddleware(username, passwordHash, "")
 	return NewHandler(db, store, authMW, nil, nil, nil, "", nil, nil)
-}
-
-// --- HLS streaming endpoints ---
-
-func (h *Handler) handleHLSStream(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-
-	if h.hlsMgr == nil || h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "HLS not available")
-		return
-	}
-
-	// Get camera to check protocol
-	cam, err := h.db.GetCamera(r.Context(), id)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get camera")
-		return
-	}
-	if cam == nil {
-		writeError(w, http.StatusNotFound, "camera not found")
-		return
-	}
-
-	// If stream not active, start it
-	if !h.hlsMgr.IsActive(id) {
-		rec := h.camMgr.GetRecorder(id)
-		if rec == nil {
-			writeError(w, http.StatusBadRequest, "camera recorder not running")
-			return
-		}
-
-		// Get camera config for HLS options
-		camCfg := h.camMgr.GetCameraConfig(id)
-		hlsMaxFPS := 0
-		if camCfg != nil {
-			hlsMaxFPS = camCfg.HLSMaxFPS
-		}
-
-		// Try H264 recorder first
-		if h264Rec, ok := rec.(*recorder.H264Recorder); ok {
-			sps := h264Rec.SPS()
-			pps := h264Rec.PPS()
-			if sps == nil || pps == nil {
-				writeError(w, http.StatusServiceUnavailable, "SPS/PPS not available yet, waiting for video stream")
-				return
-			}
-
-			err := h.hlsMgr.StartStream(id, sps, pps, hlsMaxFPS)
-			if err != nil {
-				if err == hls.ErrMaxStreamsReached {
-					writeError(w, http.StatusServiceUnavailable, "maximum HLS streams reached")
-				} else {
-					logger.Error("failed to start HLS stream", "camera_id", id, "error", err)
-					writeError(w, http.StatusInternalServerError, "failed to start HLS stream")
-				}
-				return
-			}
-
-			// Check if sub-stream URL is configured
-			if camCfg != nil && camCfg.SubStreamURL != "" {
-				if subErr := h.hlsMgr.StartSubStreamReader(id, camCfg.SubStreamURL, false); subErr != nil {
-					logger.Warn("failed to start HLS sub-stream reader, falling back to main stream", "camera_id", id, "error", subErr)
-					// Fall back to main stream OnHLSFrame
-					h264Rec.OnHLSFrame = func(pts int64, au [][]byte) {
-						_ = h.hlsMgr.WriteH264(id, pts, au)
-					}
-				}
-				// Sub-stream reader is running — do NOT set OnHLSFrame on recorder
-			} else {
-				h264Rec.OnHLSFrame = func(pts int64, au [][]byte) {
-					_ = h.hlsMgr.WriteH264(id, pts, au)
-				}
-			}
-		} else if h265Rec, ok := rec.(*recorder.H265Recorder); ok {
-			vps := h265Rec.VPS()
-			sps := h265Rec.SPS()
-			pps := h265Rec.PPS()
-			if vps == nil || sps == nil || pps == nil {
-				writeError(w, http.StatusServiceUnavailable, "VPS/SPS/PPS not available yet, waiting for video stream")
-				return
-			}
-
-			err := h.hlsMgr.StartStreamH265(id, vps, sps, pps, hlsMaxFPS)
-			if err != nil {
-				if err == hls.ErrMaxStreamsReached {
-					writeError(w, http.StatusServiceUnavailable, "maximum HLS streams reached")
-				} else {
-					logger.Error("failed to start HLS H265 stream", "camera_id", id, "error", err)
-					writeError(w, http.StatusInternalServerError, "failed to start HLS stream")
-				}
-				return
-			}
-
-			// Check if sub-stream URL is configured
-			if camCfg != nil && camCfg.SubStreamURL != "" {
-				if subErr := h.hlsMgr.StartSubStreamReader(id, camCfg.SubStreamURL, true); subErr != nil {
-					logger.Warn("failed to start HLS sub-stream reader, falling back to main stream", "camera_id", id, "error", subErr)
-					// Fall back to main stream OnHLSFrame
-					h265Rec.OnHLSFrame = func(pts int64, au [][]byte) {
-						_ = h.hlsMgr.WriteH265(id, pts, au)
-					}
-				}
-			} else {
-				h265Rec.OnHLSFrame = func(pts int64, au [][]byte) {
-					_ = h.hlsMgr.WriteH265(id, pts, au)
-				}
-			}
-		} else if onvifRec, ok := rec.(*recorder.ONVIFRecorder); ok {
-			// ONVIF recorder delegates to H264/H265 internally
-			delegate := onvifRec.Delegate()
-			if delegate == nil {
-				writeError(w, http.StatusServiceUnavailable, "ONVIF recorder delegate not available yet")
-				return
-			}
-			// Unwrap the delegate and handle as H264/H265
-			if h264Rec, ok := delegate.(*recorder.H264Recorder); ok {
-				sps := h264Rec.SPS()
-				pps := h264Rec.PPS()
-				if sps == nil || pps == nil {
-					writeError(w, http.StatusServiceUnavailable, "SPS/PPS not available yet, waiting for video stream")
-					return
-				}
-				err := h.hlsMgr.StartStream(id, sps, pps, hlsMaxFPS)
-				if err != nil {
-					if err == hls.ErrMaxStreamsReached {
-						writeError(w, http.StatusServiceUnavailable, "maximum HLS streams reached")
-					} else {
-						writeError(w, http.StatusInternalServerError, "failed to start HLS stream")
-					}
-					return
-				}
-				h264Rec.OnHLSFrame = func(pts int64, au [][]byte) {
-					_ = h.hlsMgr.WriteH264(id, pts, au)
-				}
-			} else if h265Rec, ok := delegate.(*recorder.H265Recorder); ok {
-				vps := h265Rec.VPS()
-				sps := h265Rec.SPS()
-				pps := h265Rec.PPS()
-				if vps == nil || sps == nil || pps == nil {
-					writeError(w, http.StatusServiceUnavailable, "VPS/SPS/PPS not available yet, waiting for video stream")
-					return
-				}
-				err := h.hlsMgr.StartStreamH265(id, vps, sps, pps, hlsMaxFPS)
-				if err != nil {
-					if err == hls.ErrMaxStreamsReached {
-						writeError(w, http.StatusServiceUnavailable, "maximum HLS streams reached")
-					} else {
-						writeError(w, http.StatusInternalServerError, "failed to start HLS stream")
-					}
-					return
-				}
-				h265Rec.OnHLSFrame = func(pts int64, au [][]byte) {
-					_ = h.hlsMgr.WriteH265(id, pts, au)
-				}
-			} else {
-				writeError(w, http.StatusBadRequest, "ONVIF recorder delegate type does not support HLS")
-				return
-			}
-		} else if provider, ok := rec.(model.HLSProvider); ok {
-			codec, sps, pps, vps := provider.CodecParams()
-			if sps == nil || pps == nil {
-				writeError(w, http.StatusServiceUnavailable, "codec params not ready yet, waiting for video stream")
-				return
-			}
-			switch codec {
-			case model.FormatH264:
-				err := h.hlsMgr.StartStream(id, sps, pps, hlsMaxFPS)
-				if err != nil {
-					if err == hls.ErrMaxStreamsReached {
-						writeError(w, http.StatusServiceUnavailable, "maximum HLS streams reached")
-					} else {
-						logger.Error("failed to start HLS stream", "camera_id", id, "error", err)
-						writeError(w, http.StatusInternalServerError, "failed to start HLS stream")
-					}
-					return
-				}
-				provider.SetOnHLSFrame(func(pts int64, au [][]byte) {
-					_ = h.hlsMgr.WriteH264(id, pts, au)
-				})
-			case model.FormatH265:
-				if vps == nil {
-					writeError(w, http.StatusServiceUnavailable, "VPS not ready yet, waiting for video stream")
-					return
-				}
-				err := h.hlsMgr.StartStreamH265(id, vps, sps, pps, hlsMaxFPS)
-				if err != nil {
-					if err == hls.ErrMaxStreamsReached {
-						writeError(w, http.StatusServiceUnavailable, "maximum HLS streams reached")
-					} else {
-						logger.Error("failed to start HLS H265 stream", "camera_id", id, "error", err)
-						writeError(w, http.StatusInternalServerError, "failed to start HLS stream")
-					}
-					return
-				}
-				provider.SetOnHLSFrame(func(pts int64, au [][]byte) {
-					_ = h.hlsMgr.WriteH265(id, pts, au)
-				})
-			default:
-				writeError(w, http.StatusBadRequest, "unsupported codec for HLS streaming")
-				return
-			}
-		} else {
-			writeError(w, http.StatusBadRequest, "camera recorder does not support HLS")
-			return
-		}
-	}
-	// Proxy to muxer handler
-	if !h.hlsMgr.Handle(id, w, r) {
-		writeError(w, http.StatusServiceUnavailable, "HLS stream not available")
-		return
-	}
-}
-
-func (h *Handler) handleStopHLSStream(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-
-	if h.hlsMgr == nil {
-		writeError(w, http.StatusInternalServerError, "HLS not available")
-		return
-	}
-
-	if !h.hlsMgr.IsActive(id) {
-		writeJSON(w, http.StatusOK, map[string]string{"status": "not active"})
-		return
-	}
-
-	h.hlsMgr.StopStream(id)
-	writeJSON(w, http.StatusOK, map[string]string{"status": "stopped"})
-}
-
-// --- Settings endpoints ---
-
-func (h *Handler) handleGetSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		writeError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"cleanup": map[string]any{
-			"retention_days":         h.config.Cleanup.RetentionDays,
-			"check_interval":         h.config.Cleanup.CheckInterval,
-			"disk_threshold_percent": h.config.Cleanup.DiskThresholdPercent,
-		},
-		"webdav": map[string]any{
-			"enabled":     h.config.WebDAV.Enabled != nil && *h.config.WebDAV.Enabled,
-			"path_prefix": h.config.WebDAV.PathPrefix,
-			"read_write":  h.config.WebDAV.ReadWrite,
-		},
-		"auth": map[string]any{
-			"username":        h.config.Auth.Username,
-			"auth_configured": h.config.Auth.PasswordHash != "" || h.config.Auth.Password != "",
-		},
-	})
-}
-
-func (h *Handler) handleUpdateSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		writeError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	var body struct {
-		Cleanup *struct {
-			RetentionDays        *int    `json:"retention_days"`
-			DiskThresholdPercent *int    `json:"disk_threshold_percent"`
-			CheckInterval        *string `json:"check_interval"`
-		} `json:"cleanup"`
-		WebDAV *struct {
-			Enabled    *bool   `json:"enabled"`
-			PathPrefix *string `json:"path_prefix"`
-			ReadWrite  *bool   `json:"read_write"`
-		} `json:"webdav"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	// Update cleanup settings
-	if body.Cleanup != nil {
-		if body.Cleanup.RetentionDays != nil {
-			if *body.Cleanup.RetentionDays < 1 {
-				writeError(w, http.StatusBadRequest, "retention_days must be >= 1")
-				return
-			}
-			h.config.Cleanup.RetentionDays = *body.Cleanup.RetentionDays
-		}
-		if body.Cleanup.DiskThresholdPercent != nil {
-			if *body.Cleanup.DiskThresholdPercent < 1 || *body.Cleanup.DiskThresholdPercent > 100 {
-				writeError(w, http.StatusBadRequest, "disk_threshold_percent must be between 1 and 100")
-				return
-			}
-			h.config.Cleanup.DiskThresholdPercent = *body.Cleanup.DiskThresholdPercent
-		}
-		if body.Cleanup.CheckInterval != nil {
-			if _, err := time.ParseDuration(*body.Cleanup.CheckInterval); err != nil {
-				writeError(w, http.StatusBadRequest, "check_interval must be a valid duration (e.g., \"30m\", \"1h\")")
-				return
-			}
-			h.config.Cleanup.CheckInterval = *body.Cleanup.CheckInterval
-		}
-	}
-
-	// Update webdav settings
-	if body.WebDAV != nil {
-		if body.WebDAV.Enabled != nil {
-			if h.config.WebDAV.Enabled == nil {
-				h.config.WebDAV.Enabled = new(bool)
-			}
-			*h.config.WebDAV.Enabled = *body.WebDAV.Enabled
-		}
-		if body.WebDAV.PathPrefix != nil {
-			h.config.WebDAV.PathPrefix = *body.WebDAV.PathPrefix
-		}
-		if body.WebDAV.ReadWrite != nil {
-			h.config.WebDAV.ReadWrite = *body.WebDAV.ReadWrite
-		}
-	}
-
-	// Persist config to disk
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save config", "error", err)
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-
-// --- Merge settings endpoints ---
-
-func (h *Handler) handleGetMergeSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		writeError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"enabled":               h.config.Merge.Enabled,
-		"check_interval":        h.config.Merge.CheckInterval,
-		"window_size":           h.config.Merge.WindowSize,
-		"batch_limit":           h.config.Merge.BatchLimit,
-		"min_segment_age":       h.config.Merge.MinSegmentAge,
-		"min_segments_to_merge": h.config.Merge.MinSegmentsToMerge,
-	})
-}
-
-func (h *Handler) handleUpdateMergeSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		writeError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	var body struct {
-		Enabled            *bool   `json:"enabled"`
-		CheckInterval      *string `json:"check_interval"`
-		WindowSize         *string `json:"window_size"`
-		BatchLimit         *int    `json:"batch_limit"`
-		MinSegmentAge      *string `json:"min_segment_age"`
-		MinSegmentsToMerge *int    `json:"min_segments_to_merge"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	if body.Enabled != nil {
-		h.config.Merge.Enabled = *body.Enabled
-	}
-	if body.CheckInterval != nil {
-		if _, err := time.ParseDuration(*body.CheckInterval); err != nil {
-			writeError(w, http.StatusBadRequest, "check_interval must be a valid duration (e.g., \"30m\", \"1h\")")
-			return
-		}
-		h.config.Merge.CheckInterval = *body.CheckInterval
-	}
-	if body.WindowSize != nil {
-		if _, err := time.ParseDuration(*body.WindowSize); err != nil {
-			writeError(w, http.StatusBadRequest, "window_size must be a valid duration (e.g., \"24h\", \"48h\")")
-			return
-		}
-		h.config.Merge.WindowSize = *body.WindowSize
-	}
-	if body.BatchLimit != nil {
-		if *body.BatchLimit < 1 {
-			writeError(w, http.StatusBadRequest, "batch_limit must be >= 1")
-			return
-		}
-		h.config.Merge.BatchLimit = *body.BatchLimit
-	}
-	if body.MinSegmentAge != nil {
-		if _, err := time.ParseDuration(*body.MinSegmentAge); err != nil {
-			writeError(w, http.StatusBadRequest, "min_segment_age must be a valid duration (e.g., \"1h\", \"6h\")")
-			return
-		}
-		h.config.Merge.MinSegmentAge = *body.MinSegmentAge
-	}
-	if body.MinSegmentsToMerge != nil {
-		if *body.MinSegmentsToMerge < 1 {
-			writeError(w, http.StatusBadRequest, "min_segments_to_merge must be >= 1")
-			return
-		}
-		h.config.Merge.MinSegmentsToMerge = *body.MinSegmentsToMerge
-	}
-
-	// Persist config to disk
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save config", "error", err)
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-
-func (h *Handler) handleUpdateCameraMergeConfig(w http.ResponseWriter, r *http.Request) {
-	if h.db == nil {
-		writeError(w, http.StatusInternalServerError, "database not available")
-		return
-	}
-
-	cameraID := chi.URLParam(r, "id")
-	if cameraID == "" {
-		writeError(w, http.StatusBadRequest, "camera ID is required")
-		return
-	}
-
-	var body struct {
-		Enabled            *bool   `json:"enabled"`
-		CheckInterval      *string `json:"check_interval"`
-		WindowSize         *string `json:"window_size"`
-		BatchLimit         *int    `json:"batch_limit"`
-		MinSegmentAge      *string `json:"min_segment_age"`
-		MinSegmentsToMerge *int    `json:"min_segments_to_merge"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	// Validate duration fields
-	for _, d := range []*string{body.CheckInterval, body.WindowSize, body.MinSegmentAge} {
-		if d != nil {
-			if _, err := time.ParseDuration(*d); err != nil {
-				writeError(w, http.StatusBadRequest, "duration fields must be valid (e.g., \"30m\", \"1h\")")
-				return
-			}
-		}
-	}
-	if body.BatchLimit != nil && *body.BatchLimit < 1 {
-		writeError(w, http.StatusBadRequest, "batch_limit must be >= 1")
-		return
-	}
-	if body.MinSegmentsToMerge != nil && *body.MinSegmentsToMerge < 1 {
-		writeError(w, http.StatusBadRequest, "min_segments_to_merge must be >= 1")
-		return
-	}
-
-	if err := h.db.UpsertCameraMerge(r.Context(), cameraID,
-		body.Enabled, body.CheckInterval, body.WindowSize, body.MinSegmentAge,
-		body.BatchLimit, body.MinSegmentsToMerge); err != nil {
-		logger.Warn("failed to update camera merge config", "error", err, "camera_id", cameraID)
-		writeError(w, http.StatusInternalServerError, "failed to update merge config")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-
-func (h *Handler) handleDeleteCameraMergeConfig(w http.ResponseWriter, r *http.Request) {
-	if h.db == nil {
-		writeError(w, http.StatusInternalServerError, "database not available")
-		return
-	}
-
-	cameraID := chi.URLParam(r, "id")
-	if cameraID == "" {
-		writeError(w, http.StatusBadRequest, "camera ID is required")
-		return
-	}
-
-	// Pass all nil to clear (revert to global defaults)
-	if err := h.db.UpsertCameraMerge(r.Context(), cameraID,
-		nil, nil, nil, nil, nil, nil); err != nil {
-		logger.Warn("failed to clear camera merge config", "error", err, "camera_id", cameraID)
-		writeError(w, http.StatusInternalServerError, "failed to clear merge config")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "cleared"})
-}
-
-// --- Merge status endpoints ---
-
-func (h *Handler) handleMergeStatus(w http.ResponseWriter, r *http.Request) {
-	if h.mergeMgr == nil {
-		writeJSON(w, http.StatusOK, map[string]any{
-			"enabled": false,
-		})
-		return
-	}
-	status := h.mergeMgr.Status()
-	writeJSON(w, http.StatusOK, map[string]any{
-		"enabled":         true,
-		"last_run_time":   status.LastRunTime,
-		"segments_merged": status.SegmentsMerged,
-		"files_created":   status.FilesCreated,
-		"error_count":     status.ErrorCount,
-	})
-}
-
-func (h *Handler) handleMergePending(w http.ResponseWriter, r *http.Request) {
-	if h.mergeMgr == nil {
-		writeJSON(w, http.StatusOK, map[string]any{
-			"enabled": false,
-			"pending": map[string]int{},
-		})
-		return
-	}
-	counts := h.mergeMgr.PendingCounts(r.Context())
-	writeJSON(w, http.StatusOK, map[string]any{
-		"enabled": true,
-		"pending": counts,
-	})
-}
-
-func (h *Handler) handleBackup(w http.ResponseWriter, r *http.Request) {
-	if h.db == nil {
-		writeError(w, http.StatusInternalServerError, "database not available")
-		return
-	}
-	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
-	if err := os.MkdirAll(backupDir, 0755); err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to create backup directory")
-		return
-	}
-	filename := fmt.Sprintf("nvr-backup-%s.db", time.Now().Format("20060102-150405"))
-	destPath := filepath.Join(backupDir, filename)
-	if err := h.db.Backup(r.Context(), destPath); err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to create backup")
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "created", "file": filename})
-}
-func (h *Handler) handleListBackups(w http.ResponseWriter, r *http.Request) {
-	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
-	entries, err := os.ReadDir(backupDir)
-	if err != nil {
-		writeJSON(w, http.StatusOK, []string{})
-		return
-	}
-	var backups []string
-	for _, e := range entries {
-		if !e.IsDir() && strings.HasSuffix(e.Name(), ".db") {
-			backups = append(backups, e.Name())
-		}
-	}
-	if backups == nil {
-		backups = []string{}
-	}
-	writeJSON(w, http.StatusOK, backups)
-}
-
-// --- Xiaomi cloud endpoints ---
-
-func (h *Handler) handleXiaomiAuth(w http.ResponseWriter, r *http.Request) {
-	if h.cloudProxy == nil {
-		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
-		return
-	}
-
-	var req struct {
-		Username string `json:"username"`
-		Password string `json:"password"`
-		Region   string `json:"region,omitempty"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if req.Username == "" || req.Password == "" {
-		writeError(w, http.StatusBadRequest, "username and password are required")
-		return
-	}
-
-	region := req.Region
-	if region == "" {
-		region = "cn"
-	}
-
-	result, verification, err := h.cloudProxy.SignIn(r.Context(), req.Username, req.Password, region)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, fmt.Sprintf("authentication failed: %v", err))
-		return
-	}
-
-	if verification != nil {
-		writeJSON(w, http.StatusAccepted, verificationToResponse(verification))
-		return
-	}
-
-	// Store token in config
-	h.saveXiaomiToken(result)
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"status":  "ok",
-		"user_id": result.UserID,
-	})
-}
-
-func (h *Handler) handleXiaomiCaptcha(w http.ResponseWriter, r *http.Request) {
-	if h.cloudProxy == nil {
-		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
-		return
-	}
-
-	var req struct {
-		SessionID   string `json:"session_id"`
-		CaptchaCode string `json:"captcha_code"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if req.SessionID == "" || req.CaptchaCode == "" {
-		writeError(w, http.StatusBadRequest, "session_id and captcha_code are required")
-		return
-	}
-
-	result, verification, err := h.cloudProxy.SubmitCaptcha(r.Context(), req.SessionID, req.CaptchaCode)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, fmt.Sprintf("captcha verification failed: %v", err))
-		return
-	}
-
-	if verification != nil {
-		writeJSON(w, http.StatusAccepted, verificationToResponse(verification))
-		return
-	}
-
-	// Store token in config
-	h.saveXiaomiToken(result)
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"status":  "ok",
-		"user_id": result.UserID,
-	})
-}
-
-func (h *Handler) handleXiaomiVerify(w http.ResponseWriter, r *http.Request) {
-	if h.cloudProxy == nil {
-		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
-		return
-	}
-
-	var req struct {
-		SessionID string `json:"session_id"`
-		Ticket    string `json:"ticket"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if req.SessionID == "" || req.Ticket == "" {
-		writeError(w, http.StatusBadRequest, "session_id and ticket are required")
-		return
-	}
-
-	result, verification, err := h.cloudProxy.SubmitVerify(r.Context(), req.SessionID, req.Ticket)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, fmt.Sprintf("verification failed: %v", err))
-		return
-	}
-
-	if verification != nil {
-		writeJSON(w, http.StatusAccepted, verificationToResponse(verification))
-		return
-	}
-
-	// Store token in config
-	h.saveXiaomiToken(result)
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"status":  "ok",
-		"user_id": result.UserID,
-	})
-}
-
-func (h *Handler) handleXiaomiDevices(w http.ResponseWriter, r *http.Request) {
-	if h.cloudProxy == nil {
-		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
-		return
-	}
-
-	// Get stored token from config
-	if h.config == nil || h.config.Xiaomi.Token == "" {
-		writeJSON(w, http.StatusOK, map[string]any{
-			"devices": []CloudDeviceInfo{},
-			"message": "not authenticated",
-		})
-		return
-	}
-
-	devices, err := h.cloudProxy.ListDevices(r.Context())
-	if err != nil {
-		if strings.Contains(err.Error(), "expired") || strings.Contains(err.Error(), "auth") {
-			writeError(w, http.StatusUnauthorized, fmt.Sprintf("session expired: %v", err))
-		} else {
-			writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to get devices: %v", err))
-		}
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"devices": devices,
-	})
-}
-
-// handleXiaomiSync syncs Xiaomi cloud device info (name, model, MAC) to NVR camera config.
-// It fetches all devices from Xiaomi cloud, matches them to existing NVR cameras by DID,
-// and updates name, model, brand, and serial_number (MAC). Returns count of synced cameras.
-func (h *Handler) handleXiaomiSync(w http.ResponseWriter, r *http.Request) {
-	if h.cloudProxy == nil {
-		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
-		return
-	}
-	if h.config == nil || h.config.Xiaomi.Token == "" {
-		writeError(w, http.StatusUnauthorized, "xiaomi cloud not authenticated")
-		return
-	}
-	if h.camMgr == nil {
-		writeError(w, http.StatusInternalServerError, "camera manager not available")
-		return
-	}
-
-	devices, err := h.cloudProxy.ListDevices(r.Context())
-	if err != nil {
-		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to get devices: %v", err))
-		return
-	}
-
-	// Build DID → CloudDeviceInfo lookup
-	deviceByDID := make(map[string]*CloudDeviceInfo, len(devices))
-	for i := range devices {
-		deviceByDID[devices[i].DID] = &devices[i]
-	}
-
-	synced := 0
-	for i := range h.config.Cameras {
-		cam := &h.config.Cameras[i]
-		if cam.Protocol != "xiaomi" {
-			continue
-		}
-
-		did := cam.DID
-		if did == "" {
-			did = extractDIDFromURL(cam.URL)
-		}
-		if did == "" {
-			continue
-		}
-
-		dev, ok := deviceByDID[did]
-		if !ok {
-			continue
-		}
-
-		updates := camera.CameraUpdate{
-			Name:         &dev.Name,
-			Model:        &dev.Model,
-			Brand:        strPtr("Xiaomi"),
-			SerialNumber: &dev.MAC,
-		}
-
-		if _, err := h.camMgr.UpdateCamera(context.Background(), cam.ID, updates); err != nil {
-			logger.Warn("failed to sync xiaomi camera", "camera_id", cam.ID, "did", did, "error", err)
-			continue
-		}
-		synced++
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"synced": synced,
-		"total":  len(deviceByDID),
-	})
-}
-
-// saveXiaomiToken persists auth result to config file.
-func (h *Handler) saveXiaomiToken(result *CloudAuthResult) {
-	if h.config == nil || result == nil {
-		return
-	}
-	h.config.Xiaomi.UserID = result.UserID
-	h.config.Xiaomi.Token = result.PassToken
-	h.config.Xiaomi.Region = result.Region
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save xiaomi config", "error", err)
-	}
-
-	// Also push to the cloud proxy so it has the latest credentials
-	if h.cloudProxy != nil {
-		_ = h.cloudProxy.SetCloudConfig(context.Background(), result.UserID, result.PassToken, result.Region)
-	}
-}
-
-// verificationToResponse converts a CloudVerificationRequired to an API response map.
-func verificationToResponse(v *CloudVerificationRequired) map[string]any {
-	resp := map[string]any{
-		"status": "verification_required",
-	}
-	if len(v.Captcha) > 0 {
-		resp["captcha"] = base64.StdEncoding.EncodeToString(v.Captcha)
-	}
-	if v.VerifyPhone != "" {
-		resp["verify_phone"] = v.VerifyPhone
-	}
-	if v.VerifyEmail != "" {
-		resp["verify_email"] = v.VerifyEmail
-	}
-	if v.CaptchaSessionID != "" {
-		resp["session_id"] = v.CaptchaSessionID
-	}
-	return resp
-}
-
-
-// protocolInfo describes a protocol for the /api/protocols endpoint.
-type protocolInfo struct {
-	ID           string          `json:"id"`
-	Label        string          `json:"label"`
-	Encodings    []string        `json:"encodings"`
-	BuiltIn      bool            `json:"built_in"`
-	Capabilities map[string]bool `json:"capabilities"`
-}
-
-func (h *Handler) handleProtocols(w http.ResponseWriter, r *http.Request) {
-	protocols := []protocolInfo{
-		{
-			ID:           "rtsp",
-			Label:        "RTSP",
-			Encodings:    []string{"h264", "h265", "mjpeg"},
-			BuiltIn:      true,
-			Capabilities: map[string]bool{"hls": true, "ptz": false, "snapshot": false, "discovery": false, "auth": true},
-		},
-		{
-			ID:           "http",
-			Label:        "HTTP JPEG",
-			Encodings:    []string{"jpeg"},
-			BuiltIn:      true,
-			Capabilities: map[string]bool{"hls": false, "ptz": false, "snapshot": true, "discovery": false, "auth": true},
-		},
-		{
-			ID:           "onvif",
-			Label:        "ONVIF",
-			Encodings:    []string{"h264", "h265", "mjpeg"},
-			BuiltIn:      true,
-			Capabilities: map[string]bool{"hls": true, "ptz": true, "snapshot": false, "discovery": true, "auth": true},
-		},
-		{
-			ID:           "xiaomi",
-			Label:        "Xiaomi",
-			Encodings:    []string{"h264", "h265"},
-			BuiltIn:      true,
-			Capabilities: map[string]bool{"hls": true, "ptz": false, "snapshot": false, "discovery": true, "auth": true},
-		},
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"protocols": protocols,
-	})
-}
-
-// --- Feature toggle endpoints ---
-
-func (h *Handler) handleGetFeatures(w http.ResponseWriter, r *http.Request) {
-	flags, err := h.db.GetFeatureFlags(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, "failed to get feature flags")
-		return
-	}
-	protocols := make(map[string]bool)
-	for k, v := range flags {
-		if strings.HasPrefix(k, "protocol.") {
-			proto := strings.TrimPrefix(k, "protocol.")
-			protocols[proto] = v
-		}
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"protocols": protocols})
-}
-
-func (h *Handler) handleUpdateFeatures(w http.ResponseWriter, r *http.Request) {
-	var body struct {
-		Protocols map[string]bool `json:"protocols"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	ctx := r.Context()
-	for proto, enabled := range body.Protocols {
-		if err := h.db.SetFeatureFlag(ctx, "protocol."+proto, enabled); err != nil {
-			logger.Warn("failed to set feature flag", "protocol", proto, "error", err)
-		}
-		if h.camMgr != nil {
-			h.camMgr.SetProtocolEnabled(proto, enabled)
-		}
-	}
-	// Return updated state
-	h.handleGetFeatures(w, r)
-}
-
-// formatUptime converts a duration to a human-readable string like "2h 15m 30s".
-func formatUptime(d time.Duration) string {
-	rounded := d.Round(time.Second)
-	h := rounded / time.Hour
-	rounded -= h * time.Hour
-	m := rounded / time.Minute
-	rounded -= m * time.Minute
-	s := rounded / time.Second
-	if h > 0 {
-		return fmt.Sprintf("%dh %dm %ds", h, m, s)
-	}
-	if m > 0 {
-		return fmt.Sprintf("%dm %ds", m, s)
-	}
-	return fmt.Sprintf("%ds", s)
-}
-
-// --- System stats helpers (Linux /proc) ---
-
-func readCPURaw() (total, idle uint64, err error) {
-	data, err := os.ReadFile("/proc/stat")
-	if err != nil {
-		return 0, 0, err
-	}
-	lines := strings.Split(string(data), "\n")
-	if len(lines) == 0 {
-		return 0, 0, fmt.Errorf("empty /proc/stat")
-	}
-	fields := strings.Fields(lines[0])
-	if len(fields) < 5 {
-		return 0, 0, fmt.Errorf("unexpected /proc/stat format")
-	}
-	for i := 1; i < len(fields); i++ {
-		v, _ := strconv.ParseUint(fields[i], 10, 64)
-		total += v
-	}
-	idle, _ = strconv.ParseUint(fields[4], 10, 64)
-	return
-}
-
-func readMemoryInfo() (total, available uint64, err error) {
-	data, err := os.ReadFile("/proc/meminfo")
-	if err != nil {
-		return 0, 0, err
-	}
-	for _, line := range strings.Split(string(data), "\n") {
-		fields := strings.Fields(line)
-		if len(fields) < 2 {
-			continue
-		}
-		v, _ := strconv.ParseUint(fields[1], 10, 64)
-		switch fields[0] {
-		case "MemTotal:":
-			total = v * 1024
-		case "MemAvailable:":
-			available = v * 1024
-		}
-	}
-	return
-}
-
-func readNetworkInfo() (bytesSent, bytesRecv uint64, err error) {
-	data, err := os.ReadFile("/proc/net/dev")
-	if err != nil {
-		return 0, 0, err
-	}
-	// Try eth0 or wlan0 first
-	for _, line := range strings.Split(string(data), "\n") {
-		trimmed := strings.TrimSpace(line)
-		if !strings.HasPrefix(trimmed, "eth0:") && !strings.HasPrefix(trimmed, "wlan0:") {
-			continue
-		}
-		parts := strings.SplitN(trimmed, ":", 2)
-		if len(parts) < 2 {
-			continue
-		}
-		fields := strings.Fields(parts[1])
-		if len(fields) < 10 {
-			continue
-		}
-		bytesRecv, _ = strconv.ParseUint(fields[0], 10, 64)
-		bytesSent, _ = strconv.ParseUint(fields[8], 10, 64)
-		return
-	}
-	// Fallback: sum all interfaces
-	for _, line := range strings.Split(string(data), "\n") {
-		trimmed := strings.TrimSpace(line)
-		if !strings.Contains(trimmed, ":") {
-			continue
-		}
-		parts := strings.SplitN(trimmed, ":", 2)
-		if len(parts) < 2 {
-			continue
-		}
-		fields := strings.Fields(parts[1])
-		if len(fields) < 10 {
-			continue
-		}
-		r, _ := strconv.ParseUint(fields[0], 10, 64)
-		s, _ := strconv.ParseUint(fields[8], 10, 64)
-		bytesRecv += r
-		bytesSent += s
-	}
-	return
-}
-
-func readProcessRSS() uint64 {
-	data, err := os.ReadFile("/proc/self/statm")
-	if err != nil {
-		return 0
-	}
-	fields := strings.Fields(string(data))
-	if len(fields) < 2 {
-		return 0
-	}
-	rssPages, _ := strconv.ParseUint(fields[1], 10, 64)
-	return rssPages * uint64(os.Getpagesize())
-}
-
-func (h *Handler) handleSystemStats(w http.ResponseWriter, r *http.Request) {
-	cpuTotal, cpuIdle, _ := readCPURaw()
-	memTotal, memAvailable, _ := readMemoryInfo()
-	netSent, netRecv, _ := readNetworkInfo()
-	processRSS := readProcessRSS()
-
-	writeJSON(w, http.StatusOK, SystemStats{
-		CPU:       CPUStats{Total: cpuTotal, Idle: cpuIdle},
-		Memory:    MemoryStats{Total: memTotal, Available: memAvailable, ProcessRSS: processRSS},
-		Network:   NetworkStats{BytesSent: netSent, BytesRecv: netRecv},
-		Uptime:    formatUptime(time.Since(appStartTime)),
-		Timestamp: time.Now().Unix(),
-	})
 }
 
 // extractDIDFromURL parses the DID from a xiaomi:// URL.
