+++ internal/api/handlers_onvif.go
@@ -1,0 +1,202 @@
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"time"
+
+	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/onvif"
+	"github.com/go-chi/chi/v5"
+)
+
+// --- ONVIF camera management endpoints ---
+
+func (h *Handler) handleONVIFCameraProfiles(w http.ResponseWriter, r *http.Request) {
+	cameraID := chi.URLParam(r, "id")
+	if cameraID == "" {
+		writeError(w, http.StatusBadRequest, "camera ID is required")
+		return
+	}
+
+	// For now, return empty profiles (actual implementation needs ONVIF client)
+	writeJSON(w, http.StatusOK, map[string]interface{}{
+		"profiles":     []interface{}{},
+		"capabilities": map[string]bool{"ptz": false, "streaming": false},
+	})
+}
+
+// --- ONVIF discovery endpoints ---
+
+func (h *Handler) handleONVIFDiscover(w http.ResponseWriter, r *http.Request) {
+	var req struct {
+		Timeout int `json:"timeout"`
+	}
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		req.Timeout = 5
+	}
+	if req.Timeout <= 0 {
+		req.Timeout = 5
+	}
+	if req.Timeout > 30 {
+		writeError(w, http.StatusBadRequest, "timeout must be between 1 and 30 seconds")
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(r.Context(), time.Duration(req.Timeout)*time.Second)
+	defer cancel()
+
+	devices, err := onvif.Discover(ctx, time.Duration(req.Timeout)*time.Second)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, fmt.Sprintf("discovery failed: %v", err))
+		return
+	}
+	if devices == nil {
+		devices = []onvif.DiscoveredDevice{}
+	}
+	writeJSON(w, http.StatusOK, map[string]interface{}{"devices": devices})
+}
+
+func (h *Handler) handleONVIFDeviceDetail(w http.ResponseWriter, r *http.Request) {
+	ip := chi.URLParam(r, "ip")
+	if ip == "" {
+		writeError(w, http.StatusBadRequest, "IP address is required")
+		return
+	}
+	if !validateIP(ip) {
+		writeError(w, http.StatusBadRequest, "invalid IP address format")
+		return
+	}
+	ctx := r.Context()
+	client := onvif.NewClient(fmt.Sprintf("http://%s/onvif/device_service", ip), "", "")
+	if err := client.Connect(ctx); err != nil {
+		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to connect to device: %v", err))
+		return
+	}
+	info, err := client.GetDeviceInformation(ctx)
+	if err != nil {
+		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to get device info: %v", err))
+		return
+	}
+	profiles, err := client.GetProfiles(ctx)
+	if err != nil {
+		profiles = nil
+	}
+	writeJSON(w, http.StatusOK, map[string]interface{}{
+		"device_info": info,
+		"profiles":    profiles,
+	})
+}
+
+func (h *Handler) requireONVIF(w http.ResponseWriter, r *http.Request) bool {
+	if h.db == nil {
+		writeError(w, http.StatusNotFound, "camera not found")
+		return false
+	}
+	cameraID := chi.URLParam(r, "id")
+	camera, err := h.db.GetCamera(r.Context(), cameraID)
+	if err != nil || camera == nil {
+		writeError(w, http.StatusNotFound, "camera not found")
+		return false
+	}
+	if camera.Protocol != "onvif" {
+		writeError(w, http.StatusBadRequest, "PTZ control is only available for ONVIF cameras")
+		return false
+	}
+	return true
+}
+
+// --- PTZ control endpoints ---
+
+func (h *Handler) handlePTZMove(w http.ResponseWriter, r *http.Request) {
+	cameraID := chi.URLParam(r, "id")
+	var req struct {
+		Mode string  `json:"mode"`
+		Pan  float64 `json:"pan"`
+		Tilt float64 `json:"tilt"`
+		Zoom float64 `json:"zoom"`
+	}
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		writeError(w, http.StatusBadRequest, "invalid request body")
+		return
+	}
+	if req.Mode != "continuous" && req.Mode != "absolute" && req.Mode != "relative" {
+		writeError(w, http.StatusBadRequest, "mode must be continuous, absolute, or relative")
+		return
+	}
+	if !h.requireONVIF(w, r) {
+		return
+	}
+	if h.camMgr == nil {
+		writeError(w, http.StatusInternalServerError, "camera manager not available")
+		return
+	}
+	ptz, err := h.camMgr.GetONVIFPTZController(r.Context(), cameraID)
+	if err != nil {
+		writeError(w, http.StatusBadRequest, err.Error())
+		return
+	}
+	vec := onvif.PTZVector{Pan: req.Pan, Tilt: req.Tilt, Zoom: req.Zoom}
+	switch req.Mode {
+	case "continuous":
+		err = ptz.ContinuousMove(r.Context(), vec)
+	case "absolute":
+		err = ptz.AbsoluteMove(r.Context(), vec)
+	case "relative":
+		err = ptz.RelativeMove(r.Context(), vec)
+	}
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, fmt.Sprintf("PTZ command failed: %v", err))
+		return
+	}
+	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
+}
+
+func (h *Handler) handlePTZStop(w http.ResponseWriter, r *http.Request) {
+	cameraID := chi.URLParam(r, "id")
+	if !h.requireONVIF(w, r) {
+		return
+	}
+	if h.camMgr == nil {
+		writeError(w, http.StatusInternalServerError, "camera manager not available")
+		return
+	}
+	ptz, err := h.camMgr.GetONVIFPTZController(r.Context(), cameraID)
+	if err != nil {
+		writeError(w, http.StatusBadRequest, err.Error())
+		return
+	}
+	if err := ptz.Stop(r.Context(), true, true); err != nil {
+		writeError(w, http.StatusInternalServerError, fmt.Sprintf("PTZ stop failed: %v", err))
+		return
+	}
+	writeJSON(w, http.StatusOK, map[string]string{"status": "stopped"})
+}
+
+func (h *Handler) handlePTZStatus(w http.ResponseWriter, r *http.Request) {
+	cameraID := chi.URLParam(r, "id")
+	if !h.requireONVIF(w, r) {
+		return
+	}
+	if h.camMgr == nil {
+		writeError(w, http.StatusInternalServerError, "camera manager not available")
+		return
+	}
+	ptz, err := h.camMgr.GetONVIFPTZController(r.Context(), cameraID)
+	if err != nil {
+		writeError(w, http.StatusBadRequest, err.Error())
+		return
+	}
+	pos, moving, err := ptz.GetStatus(r.Context())
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, fmt.Sprintf("get PTZ status failed: %v", err))
+		return
+	}
+	writeJSON(w, http.StatusOK, map[string]interface{}{
+		"pan":    pos.Pan,
+		"tilt":   pos.Tilt,
+		"zoom":   pos.Zoom,
+		"moving": moving,
+	})
+}
