+++ internal/api/handlers_system.go
@@ -1,0 +1,574 @@
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"os"
+	"path/filepath"
+	"runtime"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/config"
+	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/middleware"
+	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/model"
+)
+
+// --- Public endpoints ---
+
+func (h *Handler) handleHealth(w http.ResponseWriter, r *http.Request) {
+	resp := HealthResponse{Checks: make(map[string]HealthCheck)}
+	hasWarning, hasError := false, false
+
+	// Database check
+	if h.db != nil {
+		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
+		defer cancel()
+		err := h.db.DB().PingContext(ctx)
+		if err != nil {
+			resp.Checks["database"] = HealthCheck{Status: "error", Message: err.Error()}
+			hasError = true
+		} else {
+			resp.Checks["database"] = HealthCheck{Status: "ok"}
+		}
+	} else {
+		resp.Checks["database"] = HealthCheck{Status: "error", Message: "database not configured"}
+		hasError = true
+	}
+
+	// Storage check
+	if h.store != nil {
+		total, used, err := h.store.GetDiskUsage()
+		if err != nil {
+			resp.Checks["storage"] = HealthCheck{Status: "error", Message: err.Error()}
+			hasError = true
+		} else {
+			pct := 0
+			if total > 0 {
+				pct = int(float64(used) / float64(total) * 100)
+			}
+			msg := fmt.Sprintf("%d%% used (%d / %d bytes)", pct, used, total)
+			if pct > 95 {
+				resp.Checks["storage"] = HealthCheck{Status: "error", Message: msg}
+				hasError = true
+			} else if pct > 90 {
+				resp.Checks["storage"] = HealthCheck{Status: "warning", Message: msg}
+				hasWarning = true
+			} else {
+				resp.Checks["storage"] = HealthCheck{Status: "ok", Message: msg}
+			}
+		}
+	} else {
+		resp.Checks["storage"] = HealthCheck{Status: "error", Message: "storage not configured"}
+		hasError = true
+	}
+
+	// Goroutine check
+	numGoroutines := runtime.NumGoroutine()
+	if numGoroutines > 1000 {
+		resp.Checks["goroutines"] = HealthCheck{Status: "error", Message: fmt.Sprintf("%d goroutines (threshold: 1000)", numGoroutines)}
+		hasError = true
+	} else {
+		resp.Checks["goroutines"] = HealthCheck{Status: "ok", Message: fmt.Sprintf("%d goroutines", numGoroutines)}
+	}
+
+	// Overall status
+	switch {
+	case hasError:
+		resp.Status = "unhealthy"
+	case hasWarning:
+		resp.Status = "degraded"
+	default:
+		resp.Status = "ok"
+	}
+
+	// Uptime
+	resp.Uptime = formatUptime(time.Since(appStartTime))
+
+	writeJSON(w, http.StatusOK, resp)
+}
+
+func (h *Handler) handleReadyz(w http.ResponseWriter, r *http.Request) {
+	checks := make(map[string]HealthCheck)
+
+	// Database must be ok
+	allOK := true
+	if h.db == nil {
+		checks["database"] = HealthCheck{Status: "error", Message: "database not configured"}
+		allOK = false
+	} else {
+		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
+		defer cancel()
+		if err := h.db.DB().PingContext(ctx); err != nil {
+			checks["database"] = HealthCheck{Status: "error", Message: err.Error()}
+			allOK = false
+		} else {
+			checks["database"] = HealthCheck{Status: "ok"}
+		}
+	}
+
+	// Storage must be < 95%
+	if h.store == nil {
+		checks["storage"] = HealthCheck{Status: "error", Message: "storage not configured"}
+		allOK = false
+	} else {
+		total, used, err := h.store.GetDiskUsage()
+		if err != nil {
+			checks["storage"] = HealthCheck{Status: "error", Message: err.Error()}
+			allOK = false
+		} else {
+			pct := 0
+			if total > 0 {
+				pct = int(float64(used) / float64(total) * 100)
+			}
+			if pct >= 95 {
+				checks["storage"] = HealthCheck{Status: "error", Message: fmt.Sprintf("%d%% used (threshold: 95%%)", pct)}
+				allOK = false
+			} else {
+				checks["storage"] = HealthCheck{Status: "ok"}
+			}
+		}
+	}
+
+	// Goroutines must be < 5000
+	numGoroutines := runtime.NumGoroutine()
+	if numGoroutines >= 5000 {
+		checks["goroutines"] = HealthCheck{Status: "error", Message: fmt.Sprintf("%d goroutines (threshold: 5000)", numGoroutines)}
+		allOK = false
+	} else {
+		checks["goroutines"] = HealthCheck{Status: "ok"}
+	}
+
+	if allOK {
+		writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
+	} else {
+		writeJSON(w, http.StatusServiceUnavailable, map[string]any{"status": "not ready", "checks": checks})
+	}
+}
+
+func (h *Handler) handleLogin(w http.ResponseWriter, r *http.Request) {
+	// Validate credentials by running through the auth middleware.
+	// If auth is disabled, any request succeeds; otherwise BasicAuth is checked.
+	done := make(chan int, 1)
+
+	inner := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		done <- http.StatusOK
+	})
+	rec := &middleware.StatusRecorder{ResponseWriter: w, Status: http.StatusUnauthorized}
+	h.authMW(inner).ServeHTTP(rec, r)
+
+	select {
+	case status := <-done:
+		if status == http.StatusOK {
+			writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
+		}
+	default:
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusUnauthorized)
+		json.NewEncoder(w).Encode(map[string]string{"error": "unauthorized"})
+	}
+}
+
+func (h *Handler) handleStats(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+
+	total, used, err := h.store.GetDiskUsage()
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to get disk usage")
+		return
+	}
+
+	count, err := h.db.CountRecordings(ctx)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to count recordings")
+		return
+	}
+
+	cameras, err := h.db.ListCameras(ctx)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to count cameras")
+		return
+	}
+
+	stats := model.StorageStats{
+		TotalBytes:     total,
+		UsedBytes:      used,
+		RecordingCount: count,
+		CameraCount:    len(cameras),
+	}
+	writeJSON(w, http.StatusOK, stats)
+}
+
+func (h *Handler) handleStatsTrends(w http.ResponseWriter, r *http.Request) {
+	days := 7
+	if v := r.URL.Query().Get("days"); v != "" {
+		if n, err := strconv.Atoi(v); err == nil && n >= 1 && n <= 30 {
+			days = n
+		}
+	}
+	trends, err := h.db.GetRecordingTrends(r.Context(), days)
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to get recording trends")
+		return
+	}
+	writeJSON(w, http.StatusOK, trends)
+}
+
+// --- Settings endpoints ---
+
+func (h *Handler) handleGetSettings(w http.ResponseWriter, r *http.Request) {
+	if h.config == nil {
+		writeError(w, http.StatusInternalServerError, "config not available")
+		return
+	}
+
+	writeJSON(w, http.StatusOK, map[string]any{
+		"cleanup": map[string]any{
+			"retention_days":         h.config.Cleanup.RetentionDays,
+			"check_interval":         h.config.Cleanup.CheckInterval,
+			"disk_threshold_percent": h.config.Cleanup.DiskThresholdPercent,
+		},
+		"webdav": map[string]any{
+			"enabled":     h.config.WebDAV.Enabled != nil && *h.config.WebDAV.Enabled,
+			"path_prefix": h.config.WebDAV.PathPrefix,
+			"read_write":  h.config.WebDAV.ReadWrite,
+		},
+		"auth": map[string]any{
+			"username":        h.config.Auth.Username,
+			"auth_configured": h.config.Auth.PasswordHash != "" || h.config.Auth.Password != "",
+		},
+	})
+}
+
+func (h *Handler) handleUpdateSettings(w http.ResponseWriter, r *http.Request) {
+	if h.config == nil {
+		writeError(w, http.StatusInternalServerError, "config not available")
+		return
+	}
+
+	var body struct {
+		Cleanup *struct {
+			RetentionDays        *int    `json:"retention_days"`
+			DiskThresholdPercent *int    `json:"disk_threshold_percent"`
+			CheckInterval        *string `json:"check_interval"`
+		} `json:"cleanup"`
+		WebDAV *struct {
+			Enabled    *bool   `json:"enabled"`
+			PathPrefix *string `json:"path_prefix"`
+			ReadWrite  *bool   `json:"read_write"`
+		} `json:"webdav"`
+	}
+
+	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
+		writeError(w, http.StatusBadRequest, "invalid request body")
+		return
+	}
+
+	// Update cleanup settings
+	if body.Cleanup != nil {
+		if body.Cleanup.RetentionDays != nil {
+			if *body.Cleanup.RetentionDays < 1 {
+				writeError(w, http.StatusBadRequest, "retention_days must be >= 1")
+				return
+			}
+			h.config.Cleanup.RetentionDays = *body.Cleanup.RetentionDays
+		}
+		if body.Cleanup.DiskThresholdPercent != nil {
+			if *body.Cleanup.DiskThresholdPercent < 1 || *body.Cleanup.DiskThresholdPercent > 100 {
+				writeError(w, http.StatusBadRequest, "disk_threshold_percent must be between 1 and 100")
+				return
+			}
+			h.config.Cleanup.DiskThresholdPercent = *body.Cleanup.DiskThresholdPercent
+		}
+		if body.Cleanup.CheckInterval != nil {
+			if _, err := time.ParseDuration(*body.Cleanup.CheckInterval); err != nil {
+				writeError(w, http.StatusBadRequest, "check_interval must be a valid duration (e.g., \"30m\", \"1h\")")
+				return
+			}
+			h.config.Cleanup.CheckInterval = *body.Cleanup.CheckInterval
+		}
+	}
+
+	// Update webdav settings
+	if body.WebDAV != nil {
+		if body.WebDAV.Enabled != nil {
+			if h.config.WebDAV.Enabled == nil {
+				h.config.WebDAV.Enabled = new(bool)
+			}
+			*h.config.WebDAV.Enabled = *body.WebDAV.Enabled
+		}
+		if body.WebDAV.PathPrefix != nil {
+			h.config.WebDAV.PathPrefix = *body.WebDAV.PathPrefix
+		}
+		if body.WebDAV.ReadWrite != nil {
+			h.config.WebDAV.ReadWrite = *body.WebDAV.ReadWrite
+		}
+	}
+
+	// Persist config to disk
+	if err := config.Save(h.configPath, h.config); err != nil {
+		logger.Warn("failed to save config", "error", err)
+	}
+
+	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
+}
+
+func (h *Handler) handleBackup(w http.ResponseWriter, r *http.Request) {
+	if h.db == nil {
+		writeError(w, http.StatusInternalServerError, "database not available")
+		return
+	}
+	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
+	if err := os.MkdirAll(backupDir, 0755); err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to create backup directory")
+		return
+	}
+	filename := fmt.Sprintf("nvr-backup-%s.db", time.Now().Format("20060102-150405"))
+	destPath := filepath.Join(backupDir, filename)
+	if err := h.db.Backup(r.Context(), destPath); err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to create backup")
+		return
+	}
+	writeJSON(w, http.StatusOK, map[string]string{"status": "created", "file": filename})
+}
+
+func (h *Handler) handleListBackups(w http.ResponseWriter, r *http.Request) {
+	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
+	entries, err := os.ReadDir(backupDir)
+	if err != nil {
+		writeJSON(w, http.StatusOK, []string{})
+		return
+	}
+	var backups []string
+	for _, e := range entries {
+		if !e.IsDir() && strings.HasSuffix(e.Name(), ".db") {
+			backups = append(backups, e.Name())
+		}
+	}
+	if backups == nil {
+		backups = []string{}
+	}
+	writeJSON(w, http.StatusOK, backups)
+}
+
+// protocolInfo describes a protocol for the /api/protocols endpoint.
+type protocolInfo struct {
+	ID           string          `json:"id"`
+	Label        string          `json:"label"`
+	Encodings    []string        `json:"encodings"`
+	BuiltIn      bool            `json:"built_in"`
+	Capabilities map[string]bool `json:"capabilities"`
+}
+
+func (h *Handler) handleProtocols(w http.ResponseWriter, r *http.Request) {
+	protocols := []protocolInfo{
+		{
+			ID:           "rtsp",
+			Label:        "RTSP",
+			Encodings:    []string{"h264", "h265", "mjpeg"},
+			BuiltIn:      true,
+			Capabilities: map[string]bool{"hls": true, "ptz": false, "snapshot": false, "discovery": false, "auth": true},
+		},
+		{
+			ID:           "http",
+			Label:        "HTTP JPEG",
+			Encodings:    []string{"jpeg"},
+			BuiltIn:      true,
+			Capabilities: map[string]bool{"hls": false, "ptz": false, "snapshot": true, "discovery": false, "auth": true},
+		},
+		{
+			ID:           "onvif",
+			Label:        "ONVIF",
+			Encodings:    []string{"h264", "h265", "mjpeg"},
+			BuiltIn:      true,
+			Capabilities: map[string]bool{"hls": true, "ptz": true, "snapshot": false, "discovery": true, "auth": true},
+		},
+		{
+			ID:           "xiaomi",
+			Label:        "Xiaomi",
+			Encodings:    []string{"h264", "h265"},
+			BuiltIn:      true,
+			Capabilities: map[string]bool{"hls": true, "ptz": false, "snapshot": false, "discovery": true, "auth": true},
+		},
+	}
+
+	writeJSON(w, http.StatusOK, map[string]any{
+		"protocols": protocols,
+	})
+}
+
+// --- Feature toggle endpoints ---
+
+func (h *Handler) handleGetFeatures(w http.ResponseWriter, r *http.Request) {
+	flags, err := h.db.GetFeatureFlags(r.Context())
+	if err != nil {
+		writeError(w, http.StatusInternalServerError, "failed to get feature flags")
+		return
+	}
+	protocols := make(map[string]bool)
+	for k, v := range flags {
+		if strings.HasPrefix(k, "protocol.") {
+			proto := strings.TrimPrefix(k, "protocol.")
+			protocols[proto] = v
+		}
+	}
+	writeJSON(w, http.StatusOK, map[string]any{"protocols": protocols})
+}
+
+func (h *Handler) handleUpdateFeatures(w http.ResponseWriter, r *http.Request) {
+	var body struct {
+		Protocols map[string]bool `json:"protocols"`
+	}
+	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
+		writeError(w, http.StatusBadRequest, "invalid request body")
+		return
+	}
+	ctx := r.Context()
+	for proto, enabled := range body.Protocols {
+		if err := h.db.SetFeatureFlag(ctx, "protocol."+proto, enabled); err != nil {
+			logger.Warn("failed to set feature flag", "protocol", proto, "error", err)
+		}
+		if h.camMgr != nil {
+			h.camMgr.SetProtocolEnabled(proto, enabled)
+		}
+	}
+	// Return updated state
+	h.handleGetFeatures(w, r)
+}
+
+// formatUptime converts a duration to a human-readable string like "2h 15m 30s".
+func formatUptime(d time.Duration) string {
+	rounded := d.Round(time.Second)
+	h := rounded / time.Hour
+	rounded -= h * time.Hour
+	m := rounded / time.Minute
+	rounded -= m * time.Minute
+	s := rounded / time.Second
+	if h > 0 {
+		return fmt.Sprintf("%dh %dm %ds", h, m, s)
+	}
+	if m > 0 {
+		return fmt.Sprintf("%dm %ds", m, s)
+	}
+	return fmt.Sprintf("%ds", s)
+}
+
+// --- System stats helpers (Linux /proc) ---
+
+func readCPURaw() (total, idle uint64, err error) {
+	data, err := os.ReadFile("/proc/stat")
+	if err != nil {
+		return 0, 0, err
+	}
+	lines := strings.Split(string(data), "\n")
+	if len(lines) == 0 {
+		return 0, 0, fmt.Errorf("empty /proc/stat")
+	}
+	fields := strings.Fields(lines[0])
+	if len(fields) < 5 {
+		return 0, 0, fmt.Errorf("unexpected /proc/stat format")
+	}
+	for i := 1; i < len(fields); i++ {
+		v, _ := strconv.ParseUint(fields[i], 10, 64)
+		total += v
+	}
+	idle, _ = strconv.ParseUint(fields[4], 10, 64)
+	return
+}
+
+func readMemoryInfo() (total, available uint64, err error) {
+	data, err := os.ReadFile("/proc/meminfo")
+	if err != nil {
+		return 0, 0, err
+	}
+	for _, line := range strings.Split(string(data), "\n") {
+		fields := strings.Fields(line)
+		if len(fields) < 2 {
+			continue
+		}
+		v, _ := strconv.ParseUint(fields[1], 10, 64)
+		switch fields[0] {
+		case "MemTotal:":
+			total = v * 1024
+		case "MemAvailable:":
+			available = v * 1024
+		}
+	}
+	return
+}
+
+func readNetworkInfo() (bytesSent, bytesRecv uint64, err error) {
+	data, err := os.ReadFile("/proc/net/dev")
+	if err != nil {
+		return 0, 0, err
+	}
+	// Try eth0 or wlan0 first
+	for _, line := range strings.Split(string(data), "\n") {
+		trimmed := strings.TrimSpace(line)
+		if !strings.HasPrefix(trimmed, "eth0:") && !strings.HasPrefix(trimmed, "wlan0:") {
+			continue
+		}
+		parts := strings.SplitN(trimmed, ":", 2)
+		if len(parts) < 2 {
+			continue
+		}
+		fields := strings.Fields(parts[1])
+		if len(fields) < 10 {
+			continue
+		}
+		bytesRecv, _ = strconv.ParseUint(fields[0], 10, 64)
+		bytesSent, _ = strconv.ParseUint(fields[8], 10, 64)
+		return
+	}
+	// Fallback: sum all interfaces
+	for _, line := range strings.Split(string(data), "\n") {
+		trimmed := strings.TrimSpace(line)
+		if !strings.Contains(trimmed, ":") {
+			continue
+		}
+		parts := strings.SplitN(trimmed, ":", 2)
+		if len(parts) < 2 {
+			continue
+		}
+		fields := strings.Fields(parts[1])
+		if len(fields) < 10 {
+			continue
+		}
+		r, _ := strconv.ParseUint(fields[0], 10, 64)
+		s, _ := strconv.ParseUint(fields[8], 10, 64)
+		bytesRecv += r
+		bytesSent += s
+	}
+	return
+}
+
+func readProcessRSS() uint64 {
+	data, err := os.ReadFile("/proc/self/statm")
+	if err != nil {
+		return 0
+	}
+	fields := strings.Fields(string(data))
+	if len(fields) < 2 {
+		return 0
+	}
+	rssPages, _ := strconv.ParseUint(fields[1], 10, 64)
+	return rssPages * uint64(os.Getpagesize())
+}
+
+func (h *Handler) handleSystemStats(w http.ResponseWriter, r *http.Request) {
+	cpuTotal, cpuIdle, _ := readCPURaw()
+	memTotal, memAvailable, _ := readMemoryInfo()
+	netSent, netRecv, _ := readNetworkInfo()
+	processRSS := readProcessRSS()
+
+	writeJSON(w, http.StatusOK, SystemStats{
+		CPU:       CPUStats{Total: cpuTotal, Idle: cpuIdle},
+		Memory:    MemoryStats{Total: memTotal, Available: memAvailable, ProcessRSS: processRSS},
+		Network:   NetworkStats{BytesSent: netSent, BytesRecv: netRecv},
+		Uptime:    formatUptime(time.Since(appStartTime)),
+		Timestamp: time.Now().Unix(),
+	})
+}
